@@ -22,13 +22,13 @@ class MyApp extends StatelessWidget {
 }
 
 class MyHomePage extends StatelessWidget {
-  MyHomePage({Key key}) : super(key: key);
+  MyHomePage({Key? key}) : super(key: key);
 
   @override
   Widget build(BuildContext context) {
     return Scaffold(
       appBar: AppBar(
-        title: Text("sample"),
+        title: Text('sample'),
       ),
       body: SizedBox.expand(child: MyPageWithAudio()),
     );
@@ -42,43 +42,45 @@ class MyPageWithAudio extends StatefulWidget {
 
 class _MyPageWithAudioState extends State<MyPageWithAudio> {
   bool _play = false;
-  String _currentPosition = "";
+  String _currentPosition = '';
 
   @override
   Widget build(BuildContext context) {
     return AudioWidget.assets(
-      path: "assets/audios/country.mp3",
+      path: 'assets/audios/country.mp3',
       play: _play,
       onReadyToPlay: (total) {
         setState(() {
-          _currentPosition = "${Duration().mmSSFormat} / ${total.mmSSFormat}";
+          _currentPosition = '${Duration().mmSSFormat} / ${total.mmSSFormat}';
         });
       },
       onPositionChanged: (current, total) {
         setState(() {
-          _currentPosition = "${current.mmSSFormat} / ${total.mmSSFormat}";
+          _currentPosition = '${current.mmSSFormat} / ${total.mmSSFormat}';
         });
       },
       child: Column(
         mainAxisAlignment: MainAxisAlignment.center,
         mainAxisSize: MainAxisSize.max,
         children: <Widget>[
-          Text("Country music"),
+          Text('Country music'),
           Padding(
             padding: const EdgeInsets.all(8.0),
-            child: RaisedButton(
-              shape: CircleBorder(),
-              padding: EdgeInsets.all(14),
-              color: Theme.of(context).primaryColor,
-              child: Icon(
-                _play ? Icons.pause : Icons.play_arrow,
-                color: Colors.white,
+            child: ElevatedButton(
+              style: ElevatedButton.styleFrom(
+                shape: CircleBorder(),
+                padding: EdgeInsets.all(14),
+                primary: Theme.of(context).primaryColor,
               ),
               onPressed: () {
                 setState(() {
                   _play = !_play;
                 });
               },
+              child: Icon(
+                _play ? Icons.pause : Icons.play_arrow,
+                color: Colors.white,
+              ),
             ),
           ),
           Text(_currentPosition),
