@@ -2,7 +2,6 @@ import 'dart:async';
 import 'dart:io' as io;
 
 import 'package:assets_audio_player/assets_audio_player.dart';
-import 'package:file/file.dart';
 import 'package:file/local.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
@@ -11,21 +10,21 @@ import 'package:path_provider/path_provider.dart';
 
 void main() {
   SystemChrome.setEnabledSystemUIOverlays([]);
-  return runApp(new MyApp());
+  return runApp(MyApp());
 }
 
 class MyApp extends StatefulWidget {
   @override
-  _MyAppState createState() => new _MyAppState();
+  _MyAppState createState() => _MyAppState();
 }
 
 class _MyAppState extends State<MyApp> {
   @override
   Widget build(BuildContext context) {
-    return new MaterialApp(
-      home: new Scaffold(
+    return MaterialApp(
+      home: Scaffold(
         body: SafeArea(
-          child: new RecorderExample(),
+          child: RecorderExample(),
         ),
       ),
     );
@@ -36,15 +35,15 @@ class RecorderExample extends StatefulWidget {
   final LocalFileSystem localFileSystem;
 
   RecorderExample({localFileSystem})
-      : this.localFileSystem = localFileSystem ?? LocalFileSystem();
+      : localFileSystem = localFileSystem ?? LocalFileSystem();
 
   @override
-  State<StatefulWidget> createState() => new RecorderExampleState();
+  State<StatefulWidget> createState() => RecorderExampleState();
 }
 
 class RecorderExampleState extends State<RecorderExample> {
-  FlutterAudioRecorder _recorder;
-  Recording _current;
+  late FlutterAudioRecorder _recorder;
+  Recording? _current;
   RecordingStatus _currentStatus = RecordingStatus.Unset;
 
   @override
@@ -56,86 +55,89 @@ class RecorderExampleState extends State<RecorderExample> {
 
   @override
   Widget build(BuildContext context) {
-    return new Center(
-      child: new Padding(
-        padding: new EdgeInsets.all(8.0),
-        child: new Column(
-            mainAxisAlignment: MainAxisAlignment.spaceAround,
+    return Center(
+      child: Padding(
+        padding: EdgeInsets.all(8.0),
+        child:
+            Column(mainAxisAlignment: MainAxisAlignment.spaceAround, children: <
+                Widget>[
+          Row(
+            mainAxisAlignment: MainAxisAlignment.center,
             children: <Widget>[
-              new Row(
-                mainAxisAlignment: MainAxisAlignment.center,
-                children: <Widget>[
-                  Padding(
-                    padding: const EdgeInsets.all(8.0),
-                    child: new FlatButton(
-                      onPressed: () {
-                        switch (_currentStatus) {
-                          case RecordingStatus.Initialized:
-                            {
-                              _start();
-                              break;
-                            }
-                          case RecordingStatus.Recording:
-                            {
-                              _pause();
-                              break;
-                            }
-                          case RecordingStatus.Paused:
-                            {
-                              _resume();
-                              break;
-                            }
-                          case RecordingStatus.Stopped:
-                            {
-                              _init();
-                              break;
-                            }
-                          default:
-                            break;
+              Padding(
+                padding: const EdgeInsets.all(8.0),
+                child: TextButton(
+                  onPressed: () {
+                    switch (_currentStatus) {
+                      case RecordingStatus.Initialized:
+                        {
+                          _start();
+                          break;
                         }
-                      },
-                      child: _buildText(_currentStatus),
-                      color: Colors.lightBlue,
-                    ),
+                      case RecordingStatus.Recording:
+                        {
+                          _pause();
+                          break;
+                        }
+                      case RecordingStatus.Paused:
+                        {
+                          _resume();
+                          break;
+                        }
+                      case RecordingStatus.Stopped:
+                        {
+                          _init();
+                          break;
+                        }
+                      default:
+                        break;
+                    }
+                  },
+                  style: TextButton.styleFrom(
+                    primary: Colors.lightBlue,
                   ),
-                  new FlatButton(
-                    onPressed:
-                    _currentStatus != RecordingStatus.Unset ? _stop : null,
-                    child:
-                    new Text("Stop", style: TextStyle(color: Colors.white)),
-                    color: Colors.blueAccent.withOpacity(0.5),
-                  ),
-                  SizedBox(
-                    width: 8,
-                  ),
-                  new FlatButton(
-                    onPressed: onPlayAudio,
-                    child:
-                    new Text("Play", style: TextStyle(color: Colors.white)),
-                    color: Colors.blueAccent.withOpacity(0.5),
-                  ),
-                ],
+                  child: _buildText(_currentStatus),
+                ),
               ),
-              new Text("Status : $_currentStatus"),
-              new Text('Avg Power: ${_current?.metering?.averagePower}'),
-              new Text('Peak Power: ${_current?.metering?.peakPower}'),
-              new Text("File path of the record: ${_current?.path}"),
-              new Text("Format: ${_current?.audioFormat}"),
-              new Text(
-                  "isMeteringEnabled: ${_current?.metering?.isMeteringEnabled}"),
-              new Text("Extension : ${_current?.extension}"),
-              new Text(
-                  "Audio recording duration : ${_current?.duration.toString()}")
-            ]),
+              TextButton(
+                onPressed:
+                    _currentStatus != RecordingStatus.Unset ? _stop : null,
+                style: TextButton.styleFrom(
+                  primary: Colors.blueAccent.withOpacity(0.5),
+                ),
+                child: Text('Stop', style: TextStyle(color: Colors.white)),
+              ),
+              SizedBox(
+                width: 8,
+              ),
+              TextButton(
+                onPressed: onPlayAudio,
+                style: TextButton.styleFrom(
+                  primary: Colors.blueAccent.withOpacity(0.5),
+                ),
+                child: Text('Play', style: TextStyle(color: Colors.white)),
+              ),
+            ],
+          ),
+          Text('Status : $_currentStatus'),
+          Text('Avg Power: ${_current?.metering?.averagePower}'),
+          Text('Peak Power: ${_current?.metering?.peakPower}'),
+          Text('File path of the record: ${_current?.path}'),
+          Text('Format: ${_current?.audioFormat}'),
+          Text('isMeteringEnabled: ${_current?.metering?.isMeteringEnabled}'),
+          Text('Extension : ${_current?.extension}'),
+          Text('Audio recording duration : ${_current?.duration.toString()}')
+        ]),
       ),
     );
   }
 
-  _init() async {
+  Future<void> _init() async {
     try {
-      if (await FlutterAudioRecorder.hasPermissions) {
-        String customPath = '/flutter_audio_recorder_';
-        io.Directory appDocDirectory;
+      final hasPermissions = await FlutterAudioRecorder.hasPermissions;
+      if (hasPermissions == true) {
+        var customPath = '/flutter_audio_recorder_';
+        io.Directory? appDocDirectory;
 //        io.Directory appDocDirectory = await getApplicationDocumentsDirectory();
         if (io.Platform.isIOS) {
           appDocDirectory = await getApplicationDocumentsDirectory();
@@ -143,37 +145,41 @@ class RecorderExampleState extends State<RecorderExample> {
           appDocDirectory = await getExternalStorageDirectory();
         }
 
-        // can add extension like ".mp4" ".wav" ".m4a" ".aac"
-        customPath = appDocDirectory.path +
-            customPath +
-            DateTime.now().millisecondsSinceEpoch.toString();
+        if (appDocDirectory != null) {
+          // can add extension like '.mp4' '.wav' '.m4a' '.aac'
+          customPath = appDocDirectory.path +
+              customPath +
+              DateTime.now().millisecondsSinceEpoch.toString();
 
-        // .wav <---> AudioFormat.WAV
-        // .mp4 .m4a .aac <---> AudioFormat.AAC
-        // AudioFormat is optional, if given value, will overwrite path extension when there is conflicts.
-        _recorder =
-            FlutterAudioRecorder(customPath, audioFormat: AudioFormat.WAV);
+          // .wav <---> AudioFormat.WAV
+          // .mp4 .m4a .aac <---> AudioFormat.AAC
+          // AudioFormat is optional, if given value, will overwrite path extension when there is conflicts.
+          _recorder =
+              FlutterAudioRecorder(customPath, audioFormat: AudioFormat.WAV);
 
-        await _recorder.initialized;
-        // after initialization
-        var current = await _recorder.current(channel: 0);
-        print(current);
-        // should be "Initialized", if all working fine
-        setState(() {
-          _current = current;
-          _currentStatus = current.status;
-          print(_currentStatus);
-        });
+          await _recorder.initialized;
+          // after initialization
+          var current = await _recorder.current(channel: 0);
+          print(current);
+          // should be 'Initialized', if all working fine
+          setState(() {
+            _current = current;
+            if (current?.status != null) {
+              _currentStatus = current!.status!;
+            }
+            print(_currentStatus);
+          });
+        }
       } else {
-        Scaffold.of(context).showSnackBar(
-            new SnackBar(content: new Text("You must accept permissions")));
+        ScaffoldMessenger.of(context).showSnackBar(
+            SnackBar(content: Text('You must accept permissions')));
       }
     } catch (e) {
       print(e);
     }
   }
 
-  _start() async {
+  Future<void> _start() async {
     try {
       await _recorder.start();
       var recording = await _recorder.current(channel: 0);
@@ -181,8 +187,8 @@ class RecorderExampleState extends State<RecorderExample> {
         _current = recording;
       });
 
-      const tick = const Duration(milliseconds: 50);
-      new Timer.periodic(tick, (Timer t) async {
+      const tick = Duration(milliseconds: 50);
+      Timer.periodic(tick, (Timer t) async {
         if (_currentStatus == RecordingStatus.Stopped) {
           t.cancel();
         }
@@ -191,7 +197,9 @@ class RecorderExampleState extends State<RecorderExample> {
         // print(current.status);
         setState(() {
           _current = current;
-          _currentStatus = _current.status;
+          if (current?.status != null) {
+            _currentStatus = current!.status!;
+          }
         });
       });
     } catch (e) {
@@ -199,30 +207,36 @@ class RecorderExampleState extends State<RecorderExample> {
     }
   }
 
-  _resume() async {
+  Future<void> _resume() async {
     await _recorder.resume();
     setState(() {});
   }
 
-  _pause() async {
+  Future<void> _pause() async {
     await _recorder.pause();
     setState(() {});
   }
 
-  _stop() async {
-    var result = await _recorder.stop();
-    print("Stop recording: ${result.path}");
-    print("Stop recording: ${result.duration}");
-    File file = widget.localFileSystem.file(result.path);
-    print("File length: ${await file.length()}");
-    setState(() {
-      _current = result;
-      _currentStatus = _current.status;
-    });
+  Future<void> _stop() async {
+    final result = await _recorder.stop();
+    if (result != null) {
+      if (result.path != null) {
+        print('Stop recording: ${result.path}');
+        if (result.duration != null) {
+          print('Stop recording: ${result.duration}');
+          final file = widget.localFileSystem.file(result.path);
+          print('File length: ${await file.length()}');
+        }
+        setState(() {
+          _current = result;
+          if (_current!.status != null) _currentStatus = _current!.status!;
+        });
+      }
+    }
   }
 
   Widget _buildText(RecordingStatus status) {
-    var text = "";
+    var text = '';
     switch (_currentStatus) {
       case RecordingStatus.Initialized:
         {
@@ -251,7 +265,9 @@ class RecorderExampleState extends State<RecorderExample> {
   }
 
   void onPlayAudio() async {
-    final AssetsAudioPlayer assetsAudioPlayer = AssetsAudioPlayer();
-    await assetsAudioPlayer.open( Audio.file(_current.path));
+    final assetsAudioPlayer = AssetsAudioPlayer();
+    if (_current?.path != null) {
+      await assetsAudioPlayer.open(Audio.file(_current!.path!));
+    }
   }
 }
