@@ -12,22 +12,21 @@ class MyApp extends StatefulWidget {
 
 class _MyAppState extends State {
   final audio = Audio.liveStream(
-    "http://bbcmedia.ic.llnwd.net/stream/bbcmedia_radio1_mf_p",
+    'http://bbcmedia.ic.llnwd.net/stream/bbcmedia_radio1_mf_p',
     metas: Metas(
-      title: "Online",
-      artist: "Florent Champigny",
-      album: "OnlineAlbum",
+      title: 'Online',
+      artist: 'Florent Champigny',
+      album: 'OnlineAlbum',
       image: MetasImage.network(
-          "https://image.shutterstock.com/image-vector/pop-music-text-art-colorful-600w-515538502.jpg"),
+          'https://image.shutterstock.com/image-vector/pop-music-text-art-colorful-600w-515538502.jpg'),
     ),
   );
-
 
   final AssetsAudioPlayer _assetsAudioPlayer = AssetsAudioPlayer();
 
   @override
   void initState() {
-    _assetsAudioPlayer.open(this.audio, showNotification: true);
+    _assetsAudioPlayer.open(audio, showNotification: true);
     super.initState();
   }
 
@@ -36,15 +35,14 @@ class _MyAppState extends State {
     return MaterialApp(
       home: Scaffold(
         body: Center(
-          child: RaisedButton(
-            child: Text('Click'),
+          child: ElevatedButton(
             onPressed: () {
-
-              this.audio.updateMetas(
-                title: "Online 2",
-                artist: "My artist",
+              audio.updateMetas(
+                title: 'Online 2',
+                artist: 'My artist',
               );
             },
+            child: Text('Click'),
           ),
         ),
       ),
