@@ -13,7 +13,7 @@ class ForwardRewindSelector extends StatelessWidget {
         crossAxisAlignment: WrapCrossAlignment.center,
         children: <Widget>[
           Text(
-            "Forward/Rewind ",
+            'Forward/Rewind ',
             style: TextStyle(fontWeight: FontWeight.bold),
           ),
           _button(-2),
@@ -27,22 +27,22 @@ class ForwardRewindSelector extends StatelessWidget {
     return Padding(
       padding: const EdgeInsets.all(8.0),
       child: NeumorphicRadio(
-        groupValue: this.speed,
+        groupValue: speed,
         padding: EdgeInsets.all(12.0),
         value: value,
         style: NeumorphicRadioStyle(
           boxShape: NeumorphicBoxShape.circle(),
         ),
-        child: Text("x$value"),
-        onChanged: (v) {
-          this.onChange(v);
+        onChanged: (double? v) {
+          if (v != null) onChange(v);
         },
+        child: Text('x$value'),
       ),
     );
   }
 
   const ForwardRewindSelector({
-    @required this.speed,
-    @required this.onChange,
+    required this.speed,
+    required this.onChange,
   });
 }
