@@ -8,14 +8,14 @@ class PlayingControlsSmall extends StatelessWidget {
   final bool isPlaying;
   final LoopMode loopMode;
   final Function() onPlay;
-  final Function() onStop;
-  final Function() toggleLoop;
+  final Function()? onStop;
+  final Function()? toggleLoop;
 
   PlayingControlsSmall({
-    @required this.isPlaying,
-    @required this.loopMode,
+    required this.isPlaying,
+    required this.loopMode,
     this.toggleLoop,
-    @required this.onPlay,
+    required this.onPlay,
     this.onStop,
   });
 
@@ -31,14 +31,14 @@ class PlayingControlsSmall extends StatelessWidget {
           ),
           padding: EdgeInsets.all(12),
           value: LoopMode.playlist,
-          groupValue: this.loopMode,
+          groupValue: loopMode,
+          onChanged: (newValue) {
+            if (toggleLoop != null) toggleLoop!();
+          },
           child: Icon(
             Icons.loop,
             size: 18,
           ),
-          onChanged: (newValue) {
-            toggleLoop();
-          },
         ),
         SizedBox(
           width: 12,
@@ -48,7 +48,7 @@ class PlayingControlsSmall extends StatelessWidget {
             boxShape: NeumorphicBoxShape.circle(),
           ),
           padding: EdgeInsets.all(16),
-          onPressed: this.onPlay,
+          onPressed: onPlay,
           child: Icon(
             isPlaying
                 ? AssetAudioPlayerIcons.pause
@@ -56,13 +56,13 @@ class PlayingControlsSmall extends StatelessWidget {
             size: 32,
           ),
         ),
-        if(onStop != null)
+        if (onStop != null)
           NeumorphicButton(
             style: NeumorphicStyle(
               boxShape: NeumorphicBoxShape.circle(),
             ),
             padding: EdgeInsets.all(16),
-            onPressed: this.onPlay,
+            onPressed: onPlay,
             child: Icon(
               AssetAudioPlayerIcons.stop,
               size: 32,
