@@ -31,26 +31,28 @@ export 'loop.dart';
 export 'errors.dart';
 export 'PhoneStrategy.dart';
 
-const _DEFAULT_AUTO_START = true;
-const _DEFAULT_RESPECT_SILENT_MODE = false;
-const _DEFAULT_SHOW_NOTIFICATION = false;
-const _DEFAULT_PLAY_IN_BACKGROUND = PlayInBackground.enabled;
-const _DEFAULT_PLAYER = "DEFAULT_PLAYER";
+const bool _DEFAULT_AUTO_START = true;
+const bool _DEFAULT_RESPECT_SILENT_MODE = false;
+const bool _DEFAULT_SHOW_NOTIFICATION = false;
+const PlayInBackground _DEFAULT_PLAY_IN_BACKGROUND = PlayInBackground.enabled;
+const HeadPhoneStrategy _DEFAULT_HEADPHONE_STRATEGY = HeadPhoneStrategy.none;
+const LoopMode _DEFAULT_LOOP_MODE = LoopMode.none;
+const String _DEFAULT_PLAYER = 'DEFAULT_PLAYER';
 
-const METHOD_POSITION = "player.position";
-const METHOD_VOLUME = "player.volume";
-const METHOD_FINISHED = "player.finished";
-const METHOD_IS_PLAYING = "player.isPlaying";
-const METHOD_IS_BUFFERING = "player.isBuffering";
-const METHOD_CURRENT = "player.current";
-const METHOD_FORWARD_REWIND_SPEED = "player.forwardRewind";
-const METHOD_NOTIFICATION_NEXT = "player.next";
-const METHOD_NOTIFICATION_PREV = "player.prev";
-const METHOD_NOTIFICATION_STOP = "player.stop";
-const METHOD_NOTIFICATION_PLAY_OR_PAUSE = "player.playOrPause";
-const METHOD_PLAY_SPEED = "player.playSpeed";
-const METHOD_ERROR = "player.error";
-const METHOD_AUDIO_SESSION_ID = "player.audioSessionId";
+const String METHOD_POSITION = 'player.position';
+const String METHOD_VOLUME = 'player.volume';
+const String METHOD_FINISHED = 'player.finished';
+const String METHOD_IS_PLAYING = 'player.isPlaying';
+const String METHOD_IS_BUFFERING = 'player.isBuffering';
+const String METHOD_CURRENT = 'player.current';
+const String METHOD_FORWARD_REWIND_SPEED = 'player.forwardRewind';
+const String METHOD_NOTIFICATION_NEXT = 'player.next';
+const String METHOD_NOTIFICATION_PREV = 'player.prev';
+const String METHOD_NOTIFICATION_STOP = 'player.stop';
+const String METHOD_NOTIFICATION_PLAY_OR_PAUSE = 'player.playOrPause';
+const String METHOD_PLAY_SPEED = 'player.playSpeed';
+const String METHOD_ERROR = 'player.error';
+const String METHOD_AUDIO_SESSION_ID = 'player.audioSessionId';
 
 enum PlayerState {
   play,
@@ -59,33 +61,37 @@ enum PlayerState {
 }
 
 class PlayerEditor {
+  const PlayerEditor(this.assetsAudioPlayer);
+
   final AssetsAudioPlayer assetsAudioPlayer;
   PlayerEditor._(this.assetsAudioPlayer);
 
   void onAudioRemovedAt(int index) {
-    if (index < assetsAudioPlayer._playlist.playlistIndex) {
-      assetsAudioPlayer._playlist.playlistIndex =
-          assetsAudioPlayer._playlist.playlistIndex - 1;
-    }
-    assetsAudioPlayer._updatePlaylistIndexes();
-    if (assetsAudioPlayer._playlist.playlistIndex == index) {
-      assetsAudioPlayer._openPlaylistCurrent();
+    if (assetsAudioPlayer._playlist != null) {
+      if (index < assetsAudioPlayer._playlist!.playlistIndex) {
+        assetsAudioPlayer._playlist!.playlistIndex =
+            assetsAudioPlayer._playlist!.playlistIndex - 1;
+      }
+      assetsAudioPlayer._updatePlaylistIndexes();
+      if (assetsAudioPlayer._playlist!.playlistIndex == index) {
+        assetsAudioPlayer._openPlaylistCurrent();
+      }
     }
   }
 
   void onAudioAddedAt(int index) {
     assetsAudioPlayer._updatePlaylistIndexes();
-    if (assetsAudioPlayer._playlist.playlistIndex == index) {
+    if (assetsAudioPlayer._playlist!.playlistIndex == index) {
       assetsAudioPlayer._openPlaylistCurrent();
     }
   }
 
   void onAudioReplacedAt(int index, bool keepPlayingPositionIfCurrent) {
     assetsAudioPlayer._updatePlaylistIndexes();
-    if (assetsAudioPlayer._playlist.playlistIndex == index) {
+    if (assetsAudioPlayer._playlist!.playlistIndex == index) {
       final currentPosition = assetsAudioPlayer.currentPosition.value;
       final isPlaying = assetsAudioPlayer.isPlaying.value ?? false;
-      //print("onAudioReplacedAt/ currentPosition : $currentPosition");
+      //print('onAudioReplacedAt/ currentPosition : $currentPosition');
       if (keepPlayingPositionIfCurrent && currentPosition != null) {
         assetsAudioPlayer._openPlaylistCurrent(
             seek: currentPosition, autoStart: isPlaying);
@@ -106,7 +112,7 @@ class PlayerEditor {
 ///     AssetsAudioPlayer _assetsAudioPlayer = AssetsAudioPlayer();
 ///
 ///     _assetsAudioPlayer.open(Audio(
-///         "/assets/audio/myAudio.mp3",
+///         '/assets/audio/myAudio.mp3',
 ///     ))
 ///
 /// Don't forget to declare the audio folder in your `pubspec.yaml`
@@ -115,7 +121,7 @@ class PlayerEditor {
 ///       assets:
 ///         - assets/audios/
 class AssetsAudioPlayer {
-  PlayerEditor _playerEditor;
+  PlayerEditor? _playerEditor;
 
   AssetsAudioPlayerCache _audioPlayerCache = defaultAssetsAudioPlayerCache;
 
@@ -144,13 +150,14 @@ class AssetsAudioPlayer {
         const MethodChannel('assets_audio_player_notification');
     _notificationOpenChannel.setMethodCallHandler((call) async {
       switch (call.method) {
-        case "selectNotification":
+        case 'selectNotification':
           {
             final String audioId = call.arguments;
-            __onNotificationClicked.value =
-                ClickedNotificationWrapper(ClickedNotification(
-              audioId: audioId,
-            ));
+            __onNotificationClicked.add(
+              ClickedNotificationWrapper(
+                ClickedNotification(audioId: audioId),
+              ),
+            );
             break;
           }
       }
@@ -161,8 +168,8 @@ class AssetsAudioPlayer {
   static StreamSubscription addNotificationOpenAction(
       NotificationOpenAction action) {
     return _onNotificationClicked.listen((ClickedNotificationWrapper clicked) {
-      if (action != null && !clicked.handled) {
-        bool handled = action(clicked.clickedNotification);
+      if (!clicked.handled) {
+        final handled = action(clicked.clickedNotification);
         clicked.handled = handled;
       }
     });
@@ -173,20 +180,18 @@ class AssetsAudioPlayer {
 
   /// The channel between the native and Dart
   final MethodChannel _sendChannel = const MethodChannel('assets_audio_player');
-  MethodChannel _recieveChannel;
+  late MethodChannel _recieveChannel;
 
   /// Stores opened asset audio path to use it on the `_current` BehaviorSubject (in `PlayingAudio`)
-  Audio _lastOpenedAssetsAudio;
+  Audio? _lastOpenedAssetsAudio;
 
-  _CurrentPlaylist _playlist;
+  _CurrentPlaylist? _playlist;
 
   final String id;
   final NetworkSettings networkSettings = NetworkSettings();
 
   set cachePathProvider(AssetsAudioPlayerCache newValue) {
-    if (newValue != null) {
-      _audioPlayerCache = newValue;
-    }
+    _audioPlayerCache = newValue;
   }
 
   bool _acceptUserOpen = true; //if false, user cannot call open method
@@ -195,15 +200,15 @@ class AssetsAudioPlayer {
     _init();
   }
 
-  static final Map<String, AssetsAudioPlayer> _players = Map();
+  static final Map<String, AssetsAudioPlayer> _players = {};
 
   static Map<String, AssetsAudioPlayer> allPlayers() {
     return Map.from(_players); //return a copy
   }
 
-  static AssetsAudioPlayer _getOrCreate({String id}) {
+  static AssetsAudioPlayer _getOrCreate({required String id}) {
     if (_players.containsKey(id)) {
-      return _players[id];
+      return _players[id]!;
     } else {
       final player = AssetsAudioPlayer._(id: id);
       _players[id] = player;
@@ -216,19 +221,19 @@ class AssetsAudioPlayer {
   /// empty constructor now create a new player
   factory AssetsAudioPlayer() => AssetsAudioPlayer.newPlayer();
 
-  factory AssetsAudioPlayer.withId(String id) =>
+  factory AssetsAudioPlayer.withId(String? id) =>
       _getOrCreate(id: id ?? uuid.v4());
 
   /// Create a new player for this audio, play it, and dispose it automatically
   static void playAndForget(
     Audio audio, {
-    double volume,
+    double? volume,
     bool respectSilentMode = _DEFAULT_RESPECT_SILENT_MODE,
-    Duration seek,
-    double playSpeed,
+    Duration? seek,
+    double? playSpeed,
   }) {
     final player = AssetsAudioPlayer.newPlayer();
-    StreamSubscription onFinished;
+    StreamSubscription? onFinished;
     onFinished = player.playlistFinished.listen((finished) {
       if (finished) {
         onFinished?.cancel();
@@ -245,19 +250,19 @@ class AssetsAudioPlayer {
     );
   }
 
-  ReadingPlaylist get readingPlaylist {
+  ReadingPlaylist? get readingPlaylist {
     if (_playlist == null) {
       return null;
     } else {
       return ReadingPlaylist(
-        //immutable copy
-        audios: _playlist.playlist.audios,
-        currentIndex: _playlist.playlistIndex,
+        // immutable copy
+        audios: _playlist!.playlist.audios,
+        currentIndex: _playlist!.playlistIndex,
       );
     }
   }
 
-  Playlist get playlist => _playlist?.playlist;
+  Playlist? get playlist => _playlist?.playlist;
 
   /// Then mediaplayer playing state (mutable)
   final BehaviorSubject<bool> _isPlaying = BehaviorSubject<bool>.seeded(false);
@@ -272,18 +277,22 @@ class AssetsAudioPlayer {
   ///         stream: _assetsAudioPlayer.currentPosition,
   ///         builder: (context, asyncSnapshot) {
   ///             final bool isPlaying = asyncSnapshot.data;
-  ///             return Text(isPlaying ? "Pause" : "Play");
+  ///             return Text(isPlaying ? 'Pause' : 'Play');
   ///         }),
   ValueStream<bool> get isPlaying => _isPlaying.stream;
-  String get getCurrentAudioTitle => _current.value.audio.audio.metas.title;
-  String get getCurrentAudioArtist => _current.value.audio.audio.metas.artist;
+  String get getCurrentAudioTitle =>
+      _current.value?.audio.audio.metas.title ?? '';
+  String get getCurrentAudioArtist =>
+      _current.value?.audio.audio.metas.artist ?? '';
   Map<String, dynamic> get getCurrentAudioextra =>
-      _current.value.audio.audio.metas.extra;
-  String get getCurrentAudioAlbum => _current.value.audio.audio.metas.album;
-  MetasImage get getCurrentAudioImage => _current.value.audio.audio.metas.image;
+      _current.value?.audio.audio.metas.extra ?? <String, dynamic>{};
+  String get getCurrentAudioAlbum =>
+      _current.value?.audio.audio.metas.album ?? '';
+  MetasImage? get getCurrentAudioImage =>
+      _current.value?.audio.audio.metas.image;
 
-  ///represent the android session id
-  ///does nothing on others platforms
+  /// represent the android session id
+  /// does nothing on others platforms
   final BehaviorSubject<int> _audioSessionId = BehaviorSubject<int>();
 
   ValueStream<int> get audioSessionId => _audioSessionId.stream;
@@ -294,7 +303,7 @@ class AssetsAudioPlayer {
   ValueStream<PlayerState> get playerState => _playerState.stream;
 
   /// Then mediaplayer playing audio (mutable)
-  final BehaviorSubject<Playing> _current = BehaviorSubject();
+  final BehaviorSubject<Playing?> _current = BehaviorSubject();
 
   /// The current playing audio, filled with the total song duration
   /// Exposes a PlayingAudio
@@ -308,10 +317,10 @@ class AssetsAudioPlayer {
   ///         final songDuration = playing.audio.duration;
   ///     })
   ///
-  ValueStream<Playing> get current => _current.stream;
+  ValueStream<Playing?> get current => _current.stream;
 
-  Stream<PlayingAudio> get onReadyToPlay =>
-      current.map((playing) => playing?.audio); //another comprehensible name
+  Stream<PlayingAudio?> get onReadyToPlay =>
+      current.map((playing) => playing?.audio); // another comprehensible name
 
   /// Called when the the complete playlist finished to play (mutable)
   final BehaviorSubject<bool> _playlistFinished =
@@ -370,12 +379,12 @@ class AssetsAudioPlayer {
   ///         stream: _assetsAudioPlayer.volume,
   ///         builder: (context, asyncSnapshot) {
   ///             final double volume = asyncSnapshot.data;
-  ///             return Text("volume: ${volume.toString()});
+  ///             return Text('volume: ${volume.toString()});
   ///         }),
   ValueStream<double> get volume => _volume.stream;
 
   final BehaviorSubject<LoopMode> _loopMode =
-      BehaviorSubject<LoopMode>.seeded(LoopMode.none);
+      BehaviorSubject<LoopMode>.seeded(_DEFAULT_LOOP_MODE);
   final BehaviorSubject<bool> _shuffle = BehaviorSubject<bool>.seeded(false);
 
   /// Called when the looping state changes
@@ -393,23 +402,23 @@ class AssetsAudioPlayer {
   ValueStream<RealtimePlayingInfos> get realtimePlayingInfos =>
       _realtimePlayingInfos.stream;
 
-  AssetsAudioPlayerErrorHandler
-      onErrorDo; //custom error Handler, default value in "_init"
+  AssetsAudioPlayerErrorHandler?
+      onErrorDo; // custom error Handler, default value in '_init'
 
-  BehaviorSubject<double> _playSpeed = BehaviorSubject.seeded(1.0);
+  final BehaviorSubject<double> _playSpeed = BehaviorSubject.seeded(1.0);
 
   ValueStream<double> get playSpeed => _playSpeed.stream;
 
-  BehaviorSubject<double> _forwardRewindSpeed = BehaviorSubject.seeded(0);
+  final BehaviorSubject<double> _forwardRewindSpeed = BehaviorSubject.seeded(0);
 
   ValueStream<double> get forwardRewindSpeed => _forwardRewindSpeed.stream;
 
-  Duration _lastSeek;
+  Duration? _lastSeek;
 
   /// returns the looping state : true -> looping, false -> not looping
-  LoopMode get currentLoopMode => _loopMode.value;
+  LoopMode? get currentLoopMode => _loopMode.value;
 
-  bool get shuffle => _shuffle.value;
+  bool get shuffle => _shuffle.value ?? false;
 
   bool _stopped = false;
 
@@ -423,42 +432,46 @@ class AssetsAudioPlayer {
     _showNotification = newValue;
 
     /* await */ _sendChannel.invokeMethod(
-        'showNotification', {"id": this.id, "show": _showNotification});
+        'showNotification', {'id': id, 'show': _showNotification});
   }
 
   Future<void> setLoopMode(LoopMode value) async {
-    _playlist.loopMode = value;
-    _loopMode.value = value;
-    if (_playlist.isSingleAudio || value == LoopMode.single) {
-      _loopSingleAudio(value != LoopMode.none);
-    } else {
-      _loopSingleAudio(false);
+    if (_playlist != null) {
+      _playlist!.loopMode = value;
+      _loopMode.add(value);
+      if (_playlist!.isSingleAudio || value == LoopMode.single) {
+        await _loopSingleAudio(value != LoopMode.none);
+      } else {
+        await _loopSingleAudio(false);
+      }
     }
   }
 
   /// assign the shuffling state : true -> shuffling, false -> not shuffling
   set shuffle(value) {
-    _shuffle.value = value;
+    _shuffle.add(value);
   }
 
   /// toggle the looping state
   /// if it was looping -> stops this
   /// if it was'nt looping -> now it is
   Future<void> toggleLoop() async {
-    final currentMode = loopMode.value;
-    if (_playlist.isSingleAudio) {
-      if (currentMode == LoopMode.none) {
-        await setLoopMode(LoopMode.single);
+    if (_playlist != null) {
+      final currentMode = loopMode.value;
+      if (_playlist!.isSingleAudio) {
+        if (currentMode == LoopMode.none) {
+          await setLoopMode(LoopMode.single);
+        } else {
+          await setLoopMode(LoopMode.none);
+        }
       } else {
-        await setLoopMode(LoopMode.none);
-      }
-    } else {
-      if (currentMode == LoopMode.none) {
-        await setLoopMode(LoopMode.playlist);
-      } else if (currentMode == LoopMode.playlist) {
-        await setLoopMode(LoopMode.single);
-      } else {
-        await setLoopMode(LoopMode.none);
+        if (currentMode == LoopMode.none) {
+          await setLoopMode(LoopMode.playlist);
+        } else if (currentMode == LoopMode.playlist) {
+          await setLoopMode(LoopMode.single);
+        } else {
+          await setLoopMode(LoopMode.none);
+        }
       }
     }
   }
@@ -475,34 +488,38 @@ class AssetsAudioPlayer {
   Future<void> dispose() async {
     await stop();
 
-    playlist?.removeCurrentlyOpenedIn(_playerEditor);
+    if (_playerEditor != null) {
+      playlist?.removeCurrentlyOpenedIn(_playerEditor!);
+    }
 
-    _currentPosition.close();
-    _isPlaying.close();
-    _volume.close();
-    _playlistFinished.close();
-    _current.close();
-    _playlistAudioFinished.close();
-    _audioSessionId.close();
-    _loopMode.close();
-    _shuffle.close();
-    _cacheDownloadInfos.close();
-    _playSpeed.close();
-    _playerState.close();
-    _isBuffering.close();
-    _forwardRewindSpeed.close();
-    _realtimePlayingInfos.close();
-    _realTimeSubscription?.cancel();
-    _players.remove(this.id);
+    await _currentPosition.close();
+    await _isPlaying.close();
+    await _volume.close();
+    await _playlistFinished.close();
+    await _current.close();
+    await _playlistAudioFinished.close();
+    await _audioSessionId.close();
+    await _loopMode.close();
+    await _shuffle.close();
+    await _cacheDownloadInfos.close();
+    await _playSpeed.close();
+    await _playerState.close();
+    await _isBuffering.close();
+    await _forwardRewindSpeed.close();
+    await _realtimePlayingInfos.close();
+    await _realTimeSubscription?.cancel();
+    _players.remove(id);
 
     _playerEditor = null;
 
-    WidgetsBinding.instance.removeObserver(_lifecycleObserver);
-    _lifecycleObserver = null;
+    if (_lifecycleObserver != null) {
+      WidgetsBinding.instance?.removeObserver(_lifecycleObserver!);
+      _lifecycleObserver = null;
+    }
   }
 
-  _init() {
-    //default action, can be overriden using player.onErrorDo = (error, player) { ACTION };
+  void _init() {
+    // default action, can be overriden using player.onErrorDo = (error, player) { ACTION };
     onErrorDo = (errorHandler) {
       print(errorHandler.error.message);
       errorHandler.player.stop();
@@ -512,13 +529,13 @@ class AssetsAudioPlayer {
 
     _recieveChannel = MethodChannel('assets_audio_player/$id');
     _recieveChannel.setMethodCallHandler((MethodCall call) async {
-      //print("received call ${call.method} with arguments ${call.arguments}");
+      // print('received call ${call.method} with arguments ${call.arguments}');
       switch (call.method) {
         case 'log':
-          print("log: " + call.arguments);
+          print('log: ' + call.arguments);
           break;
         case METHOD_FINISHED:
-          _onFinished(call.arguments);
+          await _onFinished(call.arguments);
           break;
         case METHOD_NOTIFICATION_NEXT:
           _notificationNext();
@@ -529,7 +546,7 @@ class AssetsAudioPlayer {
         case METHOD_NOTIFICATION_STOP:
           _notificationStop();
           break;
-        case METHOD_NOTIFICATION_PLAY_OR_PAUSE: //eg: from notification
+        case METHOD_NOTIFICATION_PLAY_OR_PAUSE: // eg: from notification
           _notificationPlayPause();
           break;
         case METHOD_ERROR:
@@ -537,12 +554,12 @@ class AssetsAudioPlayer {
           break;
         case METHOD_AUDIO_SESSION_ID:
           if (call.arguments != null) {
-            _audioSessionId.value = call.arguments;
+            _audioSessionId.add(call.arguments);
           }
           break;
         case METHOD_CURRENT:
           if (call.arguments == null) {
-            final current = this._current.value;
+            final current = _current.value;
             if (current != null) {
               final finishedPlay = Playing(
                 audio: current.audio,
@@ -552,30 +569,32 @@ class AssetsAudioPlayer {
               );
               _playlistAudioFinished.add(finishedPlay);
             }
-            _playlistFinished.value = true;
-            _current.value = null;
-            _playerState.value = PlayerState.stop;
+            _playlistFinished.add(true);
+            _current.add(null);
+            _playerState.add(PlayerState.stop);
           } else {
             final totalDurationMs =
-                _toDuration(call.arguments["totalDurationMs"]);
+                _toDuration(call.arguments['totalDurationMs']);
 
-            final playingAudio = PlayingAudio(
-              audio: _lastOpenedAssetsAudio,
-              duration: totalDurationMs,
-            );
-
-            if (_playlist != null) {
-              final current = Playing(
-                audio: playingAudio,
-                index: _playlist.playlistIndex,
-                hasNext: _playlist.hasNext(),
-                playlist: ReadingPlaylist(
-                    audios: _playlist.playlist.audios,
-                    currentIndex: _playlist.playlistIndex,
-                    nextIndex: _playlist.nextIndex(),
-                    previousIndex: _playlist.previousIndex()),
+            if (_lastOpenedAssetsAudio != null) {
+              final playingAudio = PlayingAudio(
+                audio: _lastOpenedAssetsAudio!,
+                duration: totalDurationMs,
               );
-              _current.value = current;
+
+              if (_playlist != null) {
+                final current = Playing(
+                  audio: playingAudio,
+                  index: _playlist!.playlistIndex,
+                  hasNext: _playlist!.hasNext(),
+                  playlist: ReadingPlaylist(
+                      audios: _playlist!.playlist.audios,
+                      currentIndex: _playlist!.playlistIndex,
+                      nextIndex: _playlist!.nextIndex(),
+                      previousIndex: _playlist!.previousIndex()),
+                );
+                _current.add(current);
+              }
             }
           }
           break;
@@ -585,22 +604,22 @@ class AssetsAudioPlayer {
           break;
         case METHOD_IS_PLAYING:
           final bool playing = call.arguments;
-          _isPlaying.value = playing;
-          _playerState.value = playing ? PlayerState.play : PlayerState.pause;
+          _isPlaying.add(playing);
+          _playerState.add(playing ? PlayerState.play : PlayerState.pause);
           break;
         case METHOD_VOLUME:
-          _volume.value = call.arguments;
+          _volume.add(call.arguments);
           break;
         case METHOD_IS_BUFFERING:
-          _isBuffering.value = call.arguments;
+          _isBuffering.add(call.arguments);
           break;
         case METHOD_PLAY_SPEED:
-          _playSpeed.value = call.arguments;
+          _playSpeed.add(call.arguments);
           break;
         case METHOD_FORWARD_REWIND_SPEED:
           final double newValue = call.arguments;
           if (_forwardRewindSpeed.value != newValue) {
-            _forwardRewindSpeed.value = newValue;
+            _forwardRewindSpeed.add(newValue);
           }
           break;
         default:
@@ -610,18 +629,19 @@ class AssetsAudioPlayer {
     _registerToAppLifecycle();
   }
 
-  StreamSubscription _realTimeSubscription;
+  StreamSubscription? _realTimeSubscription;
 
-  AppLifecycleObserver _lifecycleObserver;
+  AppLifecycleObserver? _lifecycleObserver;
 
-  bool _wasPlayingBeforeEnterBackground;
+  bool? _wasPlayingBeforeEnterBackground;
 
   /* = null */
   void _registerToAppLifecycle() {
     _lifecycleObserver = AppLifecycleObserver(onBackground: () {
       if (_playlist != null) {
-        switch (_playlist.playInBackground) {
+        switch (_playlist!.playInBackground) {
           case PlayInBackground.enabled:
+          case null:
             {
               /* do nothing */
             }
@@ -637,8 +657,9 @@ class AssetsAudioPlayer {
       }
     }, onForeground: () {
       if (_playlist != null) {
-        switch (_playlist.playInBackground) {
+        switch (_playlist!.playInBackground) {
           case PlayInBackground.enabled:
+          case null:
             {
               /* do nothing */
             }
@@ -649,31 +670,31 @@ class AssetsAudioPlayer {
             }
             break;
           case PlayInBackground.disabledRestoreOnForeground:
-            if (_wasPlayingBeforeEnterBackground != null) {
-              if (_wasPlayingBeforeEnterBackground) {
-                play();
-              } else {
-                /* do nothing, keep the pause */
-              }
+            if (_wasPlayingBeforeEnterBackground == true) {
+              play();
+            } else {
+              /* do nothing, keep the pause */
             }
             break;
         }
       }
     });
-    WidgetsBinding.instance.addObserver(_lifecycleObserver);
+    if (_lifecycleObserver != null) {
+      WidgetsBinding.instance?.addObserver(_lifecycleObserver!);
+    }
   }
 
   void _replaceRealtimeSubscription() {
     _realTimeSubscription?.cancel();
     _realTimeSubscription = null;
     _realTimeSubscription = CombineLatestStream.list<dynamic>([
-      this.volume,
-      this.isPlaying,
-      this.loopMode,
-      this.isShuffling,
-      this.current,
-      this.currentPosition,
-      this.isBuffering
+      volume,
+      isPlaying,
+      loopMode,
+      isShuffling,
+      current,
+      currentPosition,
+      isBuffering
     ])
         .map((values) => RealtimePlayingInfos(
               volume: values[0],
@@ -683,16 +704,18 @@ class AssetsAudioPlayer {
               current: values[4],
               currentPosition: values[5],
               isBuffering: values[6],
-              playerId: this.id,
+              playerId: id,
             ))
         .listen((readingInfos) {
-      this._realtimePlayingInfos.value = readingInfos;
+      _realtimePlayingInfos.add(readingInfos);
     });
   }
 
   Future<void> playlistPlayAtIndex(int index) async {
-    _playlist.moveTo(index);
-    await _openPlaylistCurrent();
+    if (_playlist != null) {
+      _playlist!.moveTo(index);
+      await _openPlaylistCurrent();
+    }
   }
 
   /// keepLoopMode:
@@ -700,21 +723,21 @@ class AssetsAudioPlayer {
   /// if false : the loopMode is .single => execute previous() will set it as .playlist
   Future<bool> previous({bool keepLoopMode = true}) async {
     if (_playlist != null) {
-      //more than 5 sec played, go back to the start of audio
+      // more than 5 sec played, go back to the start of audio
       if (_currentPosition.value != null &&
-          _currentPosition.value.inSeconds >= 5) {
+          _currentPosition.value!.inSeconds >= 5) {
         await seek(Duration.zero, force: true);
-      } else if (_playlist.hasPrev()) {
+      } else if (_playlist!.hasPrev()) {
         if (!keepLoopMode) {
           if (loopMode.value == LoopMode.single) {
             await setLoopMode(LoopMode.playlist);
           }
         }
-        _playlist.selectPrev();
+        _playlist!.selectPrev();
         await _openPlaylistCurrent();
         return true;
-      } else if (_playlist.playlistIndex == 0) {
-        seek(Duration.zero);
+      } else if (_playlist!.playlistIndex == 0) {
+        await seek(Duration.zero);
         return true;
       }
     }
@@ -724,30 +747,30 @@ class AssetsAudioPlayer {
 
   void _onPositionReceived(dynamic argument) {
     final oldValue = _currentPosition.value;
-    int newValue = null;
+    int? newValue;
     if (argument is int) {
-      final int value = argument;
+      final value = argument;
       newValue = value;
     } else if (argument is double) {
-      final double value = argument;
+      final value = argument;
       newValue = value.round();
     }
     if (newValue != null) {
-      _currentPosition.value = Duration(milliseconds: newValue);
+      _currentPosition.add(Duration(milliseconds: newValue));
       if (loopMode.value == LoopMode.single ||
-          (this._playlist.isSingleAudio &&
+          (_playlist?.isSingleAudio == true &&
               loopMode.value == LoopMode.playlist)) {
         final current = this.current.value;
         if (current != null) {
-          final Duration completeDuration = current.audio?.duration;
-          final oldEndReached =
-              (completeDuration.inMilliseconds - oldValue.inMilliseconds) <
-                  800; //< 800ms
+          final completeDuration = current.audio.duration;
+          final oldEndReached = (completeDuration.inMilliseconds -
+                  (oldValue?.inMilliseconds ?? 0)) <
+              800; // <800ms
           final newJustStarted = newValue < 800; //<800ms
 
-          //print("old: ${oldValue.inMilliseconds}, dur : ${completeDuration.inMilliseconds}");
+          // print('old: ${oldValue.inMilliseconds}, dur : ${completeDuration.inMilliseconds}');
           if (newJustStarted && oldEndReached) {
-            //print("loop");
+            // print('loop');
             final finishedPlay = Playing(
               audio: current.audio,
               index: current.index,
@@ -755,13 +778,13 @@ class AssetsAudioPlayer {
               playlist: current.playlist,
             );
             _playlistAudioFinished.add(finishedPlay);
-            if (_playlist.isSingleAudio) {
-              _playlistFinished.value = true;
+            if (_playlist?.isSingleAudio == true) {
+              _playlistFinished.add(true);
             }
           } else if (newJustStarted && _playlistFinished.value == true) {
-            //if was true (just finished an audio)
-            //re-set it to false
-            _playlistFinished.value = false;
+            // if was true (just finished an audio)
+            // re-set it to false
+            _playlistFinished.add(false);
           }
         }
       }
@@ -769,19 +792,19 @@ class AssetsAudioPlayer {
   }
 
   Future<void> _openPlaylistCurrent(
-      {bool autoStart = true, Duration seek}) async {
+      {bool autoStart = true, Duration? seek}) async {
     if (_playlist != null) {
       return _open(
-        _playlist.currentAudio(),
-        forcedVolume: _playlist.volume,
-        respectSilentMode: _playlist.respectSilentMode,
-        showNotification: _playlist.showNotification,
-        playSpeed: _playlist.playSpeed,
-        notificationSettings: _playlist.notificationSettings,
+        _playlist!.currentAudio(),
+        forcedVolume: _playlist!.volume,
+        respectSilentMode: _playlist!.respectSilentMode,
+        showNotification: _playlist!.showNotification,
+        playSpeed: _playlist!.playSpeed,
+        notificationSettings: _playlist!.notificationSettings,
         autoStart: autoStart,
-        loopMode: _playlist.loopMode,
-        headPhoneStrategy: _playlist.headPhoneStrategy,
-        audioFocusStrategy: _playlist.audioFocusStrategy,
+        loopMode: _playlist!.loopMode,
+        headPhoneStrategy: _playlist!.headPhoneStrategy,
+        audioFocusStrategy: _playlist!.audioFocusStrategy,
         seek: seek,
       );
     }
@@ -818,33 +841,33 @@ class AssetsAudioPlayer {
           }
         }
       }
-      if (_playlist.hasNext()) {
-        final curr = this._current.value;
+      if (_playlist!.hasNext()) {
+        final curr = _current.value;
         if (curr != null) {
           _playlistAudioFinished.add(Playing(
             audio: curr.audio,
             index: curr.index,
             hasNext: true,
-            playlist: this._current.value.playlist,
+            playlist: _current.value!.playlist,
           ));
         }
-        _playlist.selectNext();
+        _playlist!.selectNext();
         await _openPlaylistCurrent();
 
         return true;
       } else if (loopMode.value == LoopMode.playlist) {
         //last element
-        final curr = this._current.value;
+        final curr = _current.value;
         if (curr != null) {
           _playlistAudioFinished.add(Playing(
             audio: curr.audio,
             index: curr.index,
             hasNext: false,
-            playlist: this._current.value.playlist,
+            playlist: _current.value!.playlist,
           ));
         }
 
-        _playlist.returnToFirst();
+        _playlist!.returnToFirst();
         await _openPlaylistCurrent();
 
         return true;
@@ -853,17 +876,17 @@ class AssetsAudioPlayer {
         return true;
       } else if (requestByUser) {
         //last element
-        final curr = this._current.value;
+        final curr = _current.value;
         if (curr != null) {
           _playlistAudioFinished.add(Playing(
             audio: curr.audio,
             index: curr.index,
             hasNext: false,
-            playlist: this._current.value.playlist,
+            playlist: _current.value!.playlist,
           ));
         }
 
-        _playlist.returnToFirst();
+        _playlist!.returnToFirst();
         await _openPlaylistCurrent();
 
         return true;
@@ -873,19 +896,19 @@ class AssetsAudioPlayer {
   }
 
   Future<void> _onFinished(bool isFinished) async {
-    bool nextDone = await _next(stopIfLast: false, requestByUser: false);
+    final nextDone = await _next(stopIfLast: false, requestByUser: false);
     if (nextDone) {
-      _playlistFinished.value = false; //continue playing the playlist
+      _playlistFinished.add(false); // continue playing the playlist
     } else {
-      _playlistFinished.value = true; // no next elements -> finished
+      _playlistFinished.add(true); // no next elements -> finished
       await stop();
     }
   }
 
   void _handleOnError(Map args) async {
-    final String errorType = args["type"];
-    final String errorMessage = args["message"];
-    final AssetsAudioPlayerError error = AssetsAudioPlayerError(
+    final String errorType = args['type'];
+    final String errorMessage = args['message'];
+    final error = AssetsAudioPlayerError(
       errorType: parseAssetsAudioPlayerErrorType(errorType),
       message: errorMessage,
     );
@@ -903,15 +926,17 @@ class AssetsAudioPlayer {
       final errorHandler = ErrorHandler(
           player: this,
           currentPosition: currentPosition.value,
-          playlist: this._playlist?.playlist,
-          playlistIndex: this._playlist?.playlistIndex,
+          playlist: _playlist?.playlist,
+          playlistIndex: _playlist?.playlistIndex,
           error: error);
-      onErrorDo(errorHandler);
+      if (onErrorDo != null) {
+        onErrorDo!(errorHandler);
+      }
     }
   }
 
   void _updatePlaylistIndexes() {
-    _playlist.clearPlayerAudio(shuffle);
+    _playlist?.clearPlayerAudio(shuffle);
   }
 
   /// Converts a number to duration
@@ -929,7 +954,7 @@ class AssetsAudioPlayer {
 
   void _notificationPrevious() {
     if (_playlist?.notificationSettings?.customPrevAction != null) {
-      _playlist?.notificationSettings?.customPrevAction(this);
+      _playlist!.notificationSettings!.customPrevAction!(this);
     } else {
       previous();
     }
@@ -937,7 +962,7 @@ class AssetsAudioPlayer {
 
   void _notificationStop() {
     if (_playlist?.notificationSettings?.customStopAction != null) {
-      _playlist?.notificationSettings?.customStopAction(this);
+      _playlist!.notificationSettings!.customStopAction!(this);
     } else {
       stop();
     }
@@ -945,7 +970,7 @@ class AssetsAudioPlayer {
 
   void _notificationPlayPause() {
     if (_playlist?.notificationSettings?.customPlayPauseAction != null) {
-      _playlist?.notificationSettings?.customPlayPauseAction(this);
+      _playlist!.notificationSettings!.customPlayPauseAction!(this);
     } else {
       playOrPause();
     }
@@ -953,32 +978,34 @@ class AssetsAudioPlayer {
 
   void _notificationNext() {
     if (_playlist?.notificationSettings?.customNextAction != null) {
-      _playlist?.notificationSettings?.customNextAction(this);
+      _playlist!.notificationSettings!.customNextAction!(this);
     } else {
       next();
     }
   }
 
-  //private method, used in open(playlist) and open(path)
+  // private method, used in open(playlist) and open(path)
   Future<void> _open(
-    Audio audioInput, {
-    bool autoStart = _DEFAULT_AUTO_START,
-    double forcedVolume,
-    bool respectSilentMode = _DEFAULT_RESPECT_SILENT_MODE,
-    bool showNotification = _DEFAULT_SHOW_NOTIFICATION,
-    Duration seek,
-    double playSpeed,
-    LoopMode loopMode,
-    HeadPhoneStrategy headPhoneStrategy,
-    AudioFocusStrategy audioFocusStrategy,
-    NotificationSettings notificationSettings,
+    Audio? audioInput, {
+    bool? autoStart,
+    double? forcedVolume,
+    bool? respectSilentMode,
+    bool? showNotification,
+    Duration? seek,
+    double? playSpeed,
+    LoopMode? loopMode,
+    HeadPhoneStrategy? headPhoneStrategy,
+    AudioFocusStrategy? audioFocusStrategy,
+    NotificationSettings? notificationSettings,
   }) async {
-    final focusStrategy = audioFocusStrategy ?? defaultFocusStrategy;
-
+    final _autoStart = autoStart ?? _DEFAULT_AUTO_START;
+    final _loopMode = loopMode ?? _DEFAULT_LOOP_MODE;
+    final _audioFocusStrategy = audioFocusStrategy ?? defaultFocusStrategy;
     final currentAudio = _lastOpenedAssetsAudio;
+    final _headPhoneStrategy = headPhoneStrategy ?? _DEFAULT_HEADPHONE_STRATEGY;
     if (audioInput != null) {
-      _respectSilentMode = respectSilentMode;
-      _showNotification = showNotification;
+      _respectSilentMode = respectSilentMode ?? _DEFAULT_RESPECT_SILENT_MODE;
+      _showNotification = showNotification ?? _DEFAULT_SHOW_NOTIFICATION;
 
       var audio = await _handlePlatformAsset(audioInput);
       audio = await _downloadOrFetchFromCacheIfNecessary(audio);
@@ -986,31 +1013,31 @@ class AssetsAudioPlayer {
       audio.setCurrentlyOpenedIn(_playerEditor);
 
       try {
-        Map<String, dynamic> params = {
-          "id": this.id,
-          "audioType": audioTypeDescription(audio.audioType),
-          "path": audio.path,
-          "autoStart": autoStart,
-          "respectSilentMode": respectSilentMode,
-          "headPhoneStrategy": describeHeadPhoneStrategy(headPhoneStrategy),
-          "audioFocusStrategy": describeAudioFocusStrategy(focusStrategy),
-          "displayNotification": showNotification,
-          "volume": forcedVolume ?? this.volume.value ?? defaultVolume,
-          "playSpeed": playSpeed ??
+        final params = {
+          'id': id,
+          'audioType': audioTypeDescription(audio.audioType),
+          'path': audio.path,
+          'autoStart': _autoStart,
+          'respectSilentMode': _respectSilentMode,
+          'headPhoneStrategy': describeHeadPhoneStrategy(_headPhoneStrategy),
+          'audioFocusStrategy': describeAudioFocusStrategy(_audioFocusStrategy),
+          'displayNotification': _showNotification,
+          'volume': forcedVolume ?? volume.value ?? defaultVolume,
+          'playSpeed': playSpeed ??
               audio.playSpeed ??
               this.playSpeed.value ??
               defaultPlaySpeed,
         };
         if (seek != null) {
-          params["seek"] = seek.inMilliseconds.round();
+          params['seek'] = seek.inMilliseconds.round();
         }
         if (audio.package != null) {
-          params["package"] = audio.package;
+          params['package'] = audio.package.toString();
         }
         if (audio.audioType == AudioType.file ||
             audio.audioType == AudioType.network ||
             audio.audioType == AudioType.liveStream) {
-          params["networkHeaders"] =
+          params['networkHeaders'] =
               audio.networkHeaders ?? networkSettings.defaultHeaders;
         }
 
@@ -1025,12 +1052,12 @@ class AssetsAudioPlayer {
 
         await _sendChannel.invokeMethod('open', params);
 
-        await setLoopMode(loopMode);
+        await setLoopMode(_loopMode);
 
         _stopped = false;
-        _playlistFinished.value = false;
+        _playlistFinished.add(false);
       } catch (e) {
-        _lastOpenedAssetsAudio = currentAudio; //revert to the previous audio
+        _lastOpenedAssetsAudio = currentAudio; // revert to the previous audio
         _current.add(null);
         _isBuffering.add(false);
         _currentPosition.add(Duration.zero);
@@ -1046,27 +1073,25 @@ class AssetsAudioPlayer {
   }
 
   Future<void> _onAudioUpdated(Audio audio) async {
-    if (_lastOpenedAssetsAudio != null) {
-      if (_lastOpenedAssetsAudio.path == audio.path) {
-        final Map<String, dynamic> params = {
-          "id": this.id,
-          "path": audio.path,
-        };
+    if (_lastOpenedAssetsAudio?.path == audio.path) {
+      final params = {
+        'id': id,
+        'path': audio.path,
+      };
 
-        writeAudioMetasInto(params, audio.metas);
+      writeAudioMetasInto(params, audio.metas);
 
-        await _sendChannel.invokeMethod('onAudioUpdated', params);
-      }
+      await _sendChannel.invokeMethod('onAudioUpdated', params);
     }
   }
 
   Future<void> updateCurrentAudioNotification(
-      {Metas metas, bool showNotifications = true}) async {
+      {Metas? metas, bool showNotifications = true}) async {
     if (_lastOpenedAssetsAudio != null) {
-      final Map<String, dynamic> params = {
-        "id": this.id,
-        "path": _lastOpenedAssetsAudio.path,
-        "showNotification": showNotifications,
+      final params = {
+        'id': id,
+        'path': _lastOpenedAssetsAudio?.path,
+        'showNotification': showNotifications,
       };
 
       writeAudioMetasInto(params, metas);
@@ -1078,20 +1103,20 @@ class AssetsAudioPlayer {
   Future<void> _openPlaylist(
     Playlist playlist, {
     bool autoStart = _DEFAULT_AUTO_START,
-    double volume,
+    double? volume,
     bool respectSilentMode = _DEFAULT_RESPECT_SILENT_MODE,
     bool showNotification = _DEFAULT_SHOW_NOTIFICATION,
-    Duration seek,
-    double playSpeed,
-    LoopMode loopMode,
-    NotificationSettings notificationSettings,
-    PlayInBackground playInBackground = _DEFAULT_PLAY_IN_BACKGROUND,
-    HeadPhoneStrategy headPhoneStrategy = HeadPhoneStrategy.none,
-    AudioFocusStrategy audioFocusStrategy,
+    Duration? seek,
+    double? playSpeed,
+    LoopMode? loopMode,
+    NotificationSettings? notificationSettings,
+    PlayInBackground? playInBackground,
+    HeadPhoneStrategy headPhoneStrategy = _DEFAULT_HEADPHONE_STRATEGY,
+    AudioFocusStrategy? audioFocusStrategy,
   }) async {
     _lastSeek = null;
     _replaceRealtimeSubscription();
-    this._playlist = _CurrentPlaylist(
+    _playlist = _CurrentPlaylist(
       playlist: playlist,
       volume: volume,
       respectSilentMode: respectSilentMode,
@@ -1100,11 +1125,11 @@ class AssetsAudioPlayer {
       loopMode: loopMode,
       audioFocusStrategy: audioFocusStrategy ?? defaultFocusStrategy,
       notificationSettings: notificationSettings,
-      playInBackground: playInBackground,
+      playInBackground: playInBackground ?? _DEFAULT_PLAY_IN_BACKGROUND,
       headPhoneStrategy: headPhoneStrategy,
     );
     _updatePlaylistIndexes();
-    _playlist.moveTo(playlist.startIndex);
+    _playlist!.moveTo(playlist.startIndex);
 
     playlist.setCurrentlyOpenedIn(_playerEditor);
 
@@ -1120,7 +1145,7 @@ class AssetsAudioPlayer {
   ///
   ///     AssetsAudioPlayer _assetsAudioPlayer = AssetsAudioPlayer();
   ///
-  ///     _assetsAudioPlayer.open(Audio("assets/audios/song1.mp3"))
+  ///     _assetsAudioPlayer.open(Audio('assets/audios/song1.mp3'))
   ///
   /// Don't forget to declare the audio folder in your `pubspec.yaml`
   ///
@@ -1131,17 +1156,17 @@ class AssetsAudioPlayer {
   Future<void> open(
     Playable playable, {
     bool autoStart = _DEFAULT_AUTO_START,
-    double volume,
+    double? volume,
     bool respectSilentMode = _DEFAULT_RESPECT_SILENT_MODE,
     bool showNotification = _DEFAULT_SHOW_NOTIFICATION,
-    Duration seek,
-    double playSpeed,
-    NotificationSettings notificationSettings,
-    LoopMode loopMode = LoopMode.none,
+    Duration? seek,
+    double? playSpeed,
+    NotificationSettings? notificationSettings,
+    LoopMode loopMode = _DEFAULT_LOOP_MODE,
     PlayInBackground playInBackground = _DEFAULT_PLAY_IN_BACKGROUND,
-    HeadPhoneStrategy headPhoneStrategy = HeadPhoneStrategy.none,
-    AudioFocusStrategy audioFocusStrategy,
-    bool forceOpen = false, //skip the _acceptUserOpen
+    HeadPhoneStrategy headPhoneStrategy = _DEFAULT_HEADPHONE_STRATEGY,
+    AudioFocusStrategy? audioFocusStrategy,
+    bool forceOpen = false, // skip the _acceptUserOpen
   }) async {
     final focusStrategy = audioFocusStrategy ?? defaultFocusStrategy;
 
@@ -1155,10 +1180,8 @@ class AssetsAudioPlayer {
 
     try {
       _acceptUserOpen = false;
-      Playlist playlist;
-      if (playable is Playlist &&
-          playable.audios != null &&
-          playable.audios.length > 0) {
+      Playlist? playlist;
+      if (playable is Playlist && playable.audios.isNotEmpty) {
         playlist = playable;
       } else if (playable is Audio) {
         playlist = Playlist(audios: [playable]);
@@ -1184,7 +1207,7 @@ class AssetsAudioPlayer {
       _acceptUserOpen = true;
     } catch (t) {
       _acceptUserOpen = true;
-      throw t;
+      rethrow;
     }
   }
 
@@ -1195,7 +1218,7 @@ class AssetsAudioPlayer {
   ///     _assetsAudioPlayer.playOfPause();
   ///
   Future<void> playOrPause() async {
-    final bool playing = _isPlaying.value;
+    final playing = _isPlaying.value ?? true;
     if (playing) {
       await pause();
     } else {
@@ -1208,13 +1231,13 @@ class AssetsAudioPlayer {
   ///
   Future<void> play() async {
     if (_isLiveStream) {
-      //on livestream, it re-open the media to be live and not on buffer
+      // on livestream, it re-open the media to be live and not on buffer
       await _openPlaylistCurrent();
     } else {
       if (_stopped) {
         _stopped = false;
         _lastOpenedAssetsAudio = null; //to force open again
-        //open the last
+        // open the last
         _playlist?.returnToFirst();
         await _openPlaylistCurrent();
       } else {
@@ -1225,13 +1248,13 @@ class AssetsAudioPlayer {
 
   Future<void> _play() async {
     await _sendChannel.invokeMethod('play', {
-      "id": this.id,
+      'id': id,
     });
   }
 
   Future<void> _loopSingleAudio(bool loop) async {
     await _sendChannel
-        .invokeMethod('loopSingleAudio', {"id": this.id, "loop": loop});
+        .invokeMethod('loopSingleAudio', {'id': id, 'loop': loop});
   }
 
   /// Tells the media player to pause the current song
@@ -1239,12 +1262,12 @@ class AssetsAudioPlayer {
   ///
   Future<void> pause() async {
     if (_isLiveStream) {
-      //on livestream, we stop
+      // on livestream, we stop
       await _stop(removeNotification: false);
     } else {
       if (!_stopped) {
         await _sendChannel.invokeMethod('pause', {
-          "id": this.id,
+          'id': id,
         });
         _lastSeek = _currentPosition.value;
       }
@@ -1260,32 +1283,31 @@ class AssetsAudioPlayer {
     if (to != _lastSeek || force) {
       _lastSeek = to;
       await _sendChannel.invokeMethod('seek', {
-        "id": this.id,
-        "to": to.inMilliseconds.round(),
+        'id': id,
+        'to': to.inMilliseconds.round(),
       });
     }
   }
 
-  bool _wasPlayingBeforeForwardRewind;
+  bool? _wasPlayingBeforeForwardRewind;
 
   /// If positive, forward (progressively)
   /// If Negative rewind (progressively)
   /// If 0 or null, restore the playing state
   Future<void> forwardOrRewind(double speed) async {
-    if (speed == 0 || speed == null) {
-      if (_wasPlayingBeforeForwardRewind) {
+    if (speed == 0) {
+      if (_wasPlayingBeforeForwardRewind == true) {
         await play();
       } else {
         await pause();
       }
       _wasPlayingBeforeForwardRewind = null;
     } else {
-      if (_wasPlayingBeforeForwardRewind == null) {
-        _wasPlayingBeforeForwardRewind = this.isPlaying.value;
-      }
+      _wasPlayingBeforeForwardRewind ??= isPlaying.value;
+
       await _sendChannel.invokeMethod('forwardRewind', {
-        "id": this.id,
-        "speed": speed,
+        'id': id,
+        'speed': speed,
       });
     }
   }
@@ -1299,17 +1321,17 @@ class AssetsAudioPlayer {
   ///  eg: _assetsAudioPlayer.rewind(Duration(seconds: 10))
   ///
   Future<void> seekBy(Duration by) async {
-    //only if playing a song
-    final playing = this.current.value;
+    // only if playing a song
+    final playing = current.value;
     if (playing != null) {
-      final totalDuration = playing.audio?.duration ?? Duration.zero;
+      final totalDuration = playing.audio.duration;
 
       final currentPosition = this.currentPosition.value ?? Duration();
 
       if (by.inMilliseconds >= 0) {
         final nextPosition = currentPosition + by;
 
-        //don't seek more that song duration
+        // don't seek more that song duration
         final currentPositionCapped = Duration(
           milliseconds:
               min(totalDuration.inMilliseconds, nextPosition.inMilliseconds),
@@ -1317,11 +1339,11 @@ class AssetsAudioPlayer {
 
         await seek(currentPositionCapped);
       } else {
-        //only if playing a song
+        // only if playing a song
         final currentPosition = this.currentPosition.value ?? Duration();
         final nextPosition = currentPosition + by;
 
-        //don't seek less that 0
+        // don't seek less that 0
         final currentPositionCapped = Duration(
           milliseconds: max(0, nextPosition.inMilliseconds),
         );
@@ -1340,8 +1362,8 @@ class AssetsAudioPlayer {
   ///
   Future<void> setVolume(double volume) async {
     await _sendChannel.invokeMethod('volume', {
-      "id": this.id,
-      "volume": volume.clamp(minVolume, maxVolume),
+      'id': id,
+      'volume': volume.clamp(minVolume, maxVolume),
     });
   }
 
@@ -1355,7 +1377,7 @@ class AssetsAudioPlayer {
   Future<void> _stop({bool removeNotification = true}) async {
     _stopped = true;
     await _sendChannel.invokeMethod(
-        'stop', {"id": this.id, "removeNotification": removeNotification});
+        'stop', {'id': id, 'removeNotification': removeNotification});
   }
 
   /// Change the current play speed (rate) of the MediaPlayer
@@ -1369,9 +1391,8 @@ class AssetsAudioPlayer {
   ///
   Future<void> setPlaySpeed(double playSpeed) async {
     await _sendChannel.invokeMethod('playSpeed', {
-      "id": this.id,
-      "playSpeed":
-          (playSpeed ?? defaultPlaySpeed).clamp(minPlaySpeed, maxPlaySpeed),
+      'id': id,
+      'playSpeed': playSpeed.clamp(minPlaySpeed, maxPlaySpeed),
     });
   }
 
@@ -1380,7 +1401,7 @@ class AssetsAudioPlayer {
         input.audioType == AudioType.asset &&
         !kIsWeb) {
       //on macos assets are not available from native
-      final String path = await _copyToTmpMemory(
+      final path = await _copyToTmpMemory(
           package: input.package, assetSource: input.path);
       return input.copyWith(audioType: AudioType.file, path: path);
     }
@@ -1388,15 +1409,16 @@ class AssetsAudioPlayer {
   }
 
   Future<Audio> _downloadOrFetchFromCacheIfNecessary(Audio input) async {
-    return AssetsAudioPlayerCacheManager()
-        .transform(this._audioPlayerCache, input, (downloadInfos) {
-      this._cacheDownloadInfos.add(downloadInfos);
+    return AssetsAudioPlayerCacheManager().transform(_audioPlayerCache, input,
+        (downloadInfos) {
+      _cacheDownloadInfos.add(downloadInfos);
     });
   }
 
   //returns the file path
-  Future<String> _copyToTmpMemory({String package, String assetSource}) async {
-    final String fileName = "${package ?? ""}$assetSource";
+  Future<String> _copyToTmpMemory(
+      {String? package, String? assetSource}) async {
+    final fileName = '${package ?? ''}$assetSource';
     final completePath = '${(await getTemporaryDirectory()).path}/$fileName';
     final file = File(completePath);
     if (await file.exists()) {
@@ -1421,20 +1443,20 @@ class AssetsAudioPlayer {
 class _CurrentPlaylist {
   final Playlist playlist;
 
-  final double volume;
-  final bool respectSilentMode;
-  final bool showNotification;
-  LoopMode loopMode;
-  final double playSpeed;
-  final NotificationSettings notificationSettings;
-  final AudioFocusStrategy audioFocusStrategy;
-  final PlayInBackground playInBackground;
-  final HeadPhoneStrategy headPhoneStrategy;
+  final double? volume;
+  final bool? respectSilentMode;
+  final bool? showNotification;
+  LoopMode? loopMode;
+  final double? playSpeed;
+  final NotificationSettings? notificationSettings;
+  final AudioFocusStrategy? audioFocusStrategy;
+  final PlayInBackground? playInBackground;
+  final HeadPhoneStrategy? headPhoneStrategy;
 
   int playlistIndex = 0;
 
   int nextIndex() {
-    int index = indexList.indexWhere((element) => playlistIndex == element);
+    final index = indexList.indexWhere((element) => playlistIndex == element);
     if (index + 1 == indexList.length) {
       return indexList.first;
     } else {
@@ -1443,7 +1465,7 @@ class _CurrentPlaylist {
   }
 
   int previousIndex() {
-    int index = indexList.indexWhere((element) => playlistIndex == element);
+    final index = indexList.indexWhere((element) => playlistIndex == element);
     if (index == 0) {
       return indexList.last;
     } else {
@@ -1451,8 +1473,8 @@ class _CurrentPlaylist {
     }
   }
 
-  selectNext() {
-    int index = indexList.indexWhere((element) => playlistIndex == element);
+  void selectNext() {
+    var index = indexList.indexWhere((element) => playlistIndex == element);
     if (hasNext()) {
       index = index + 1;
     }
@@ -1461,13 +1483,13 @@ class _CurrentPlaylist {
 
   List<int> indexList = [];
 
-  sortAudios() {
-    for (var i = 0; i < this.playlist.audios.length; i++) {
+  void sortAudios() {
+    for (var i = 0; i < playlist.audios.length; i++) {
       indexList.add(i);
     }
   }
 
-  clearPlayerAudio(bool shuffle) {
+  void clearPlayerAudio(bool shuffle) {
     indexList.clear();
     if (shuffle) {
       shuffleAudios();
@@ -1476,16 +1498,16 @@ class _CurrentPlaylist {
     }
   }
 
-  shuffleAudios() {
-    for (var i = 0; i < this.playlist.audios.length; i++) {
-      int index = _shuffleNumbers();
+  void shuffleAudios() {
+    for (var i = 0; i < playlist.audios.length; i++) {
+      final index = _shuffleNumbers();
       indexList.add(index);
     }
   }
 
   int _shuffleNumbers() {
-    Random random = Random();
-    int index = random.nextInt(playlist.audios.length);
+    final random = Random();
+    var index = random.nextInt(playlist.audios.length);
     if (indexList.contains(index)) {
       index = _shuffleNumbers();
     }
@@ -1501,8 +1523,7 @@ class _CurrentPlaylist {
     return playlistIndex;
   }
 
-  //nullable
-  Audio audioAt({int at}) {
+  Audio? audioAt({required int at}) {
     if (at < playlist.audios.length) {
       return playlist.audios[at];
     } else {
@@ -1510,19 +1531,19 @@ class _CurrentPlaylist {
     }
   }
 
-  Audio currentAudio() {
+  Audio? currentAudio() {
     return audioAt(at: indexList[playlistIndex]);
   }
 
   bool hasNext() {
-    int index = indexList.indexWhere((element) => playlistIndex == element);
+    var index = indexList.indexWhere((element) => playlistIndex == element);
     return index + 1 < indexList.length;
   }
 
   bool get isSingleAudio => playlist.audios.length == 1;
 
   _CurrentPlaylist({
-    @required this.playlist,
+    required this.playlist,
     this.volume,
     this.respectSilentMode,
     this.showNotification,
@@ -1539,12 +1560,12 @@ class _CurrentPlaylist {
   }
 
   bool hasPrev() {
-    int index = indexList.indexWhere((element) => playlistIndex == element);
+    var index = indexList.indexWhere((element) => playlistIndex == element);
     return index > 0;
   }
 
   void selectPrev() {
-    int index = indexList.indexWhere((element) => playlistIndex == element);
+    var index = indexList.indexWhere((element) => playlistIndex == element);
     index = index - 1;
     playlistIndex = index;
     if (playlistIndex < 0) {
