@@ -1,5 +1,3 @@
-import 'package:flutter/foundation.dart';
-
 import 'assets_audio_player.dart';
 import 'player_group.dart';
 
@@ -8,16 +6,16 @@ enum AssetsAudioPlayerErrorType { Network, Player }
 class ErrorHandler {
   final AssetsAudioPlayerError error;
   final AssetsAudioPlayer player;
-  final Duration currentPosition;
-  final Playlist playlist;
-  final int playlistIndex;
+  final Duration? currentPosition;
+  final Playlist? playlist;
+  final int? playlistIndex;
 
   const ErrorHandler({
-    @required this.error,
-    @required this.player,
-    @required this.currentPosition,
-    @required this.playlist,
-    @required this.playlistIndex,
+    required this.error,
+    required this.player,
+    this.currentPosition,
+    this.playlist,
+    this.playlistIndex,
   });
 }
 
@@ -27,7 +25,7 @@ typedef AssetsAudioPlayerGroupErrorHandler = Function(
 
 AssetsAudioPlayerErrorType parseAssetsAudioPlayerErrorType(String type) {
   switch (type) {
-    case "network":
+    case 'network':
       return AssetsAudioPlayerErrorType.Network;
     default:
       return AssetsAudioPlayerErrorType.Player;
@@ -39,8 +37,8 @@ class AssetsAudioPlayerError {
   final String message;
 
   const AssetsAudioPlayerError({
-    @required this.errorType,
-    @required this.message,
+    required this.errorType,
+    required this.message,
   });
 
   @override
