@@ -5,15 +5,15 @@ import 'package:uuid/uuid.dart';
 import 'utils.dart';
 
 class Playable {
-  final Set<PlayerEditor> _currentlyOpenedIn = Set();
+  final Set<PlayerEditor> _currentlyOpenedIn = {};
 
   Set<PlayerEditor> get currentlyOpenedIn => Set.from(_currentlyOpenedIn);
 
-  void setCurrentlyOpenedIn(PlayerEditor player) {
-    _currentlyOpenedIn.add(player);
+  void setCurrentlyOpenedIn(PlayerEditor? player) {
+    if (player != null) _currentlyOpenedIn.add(player);
   }
 
-  void removeCurrentlyOpenedIn(PlayerEditor player) {
+  void removeCurrentlyOpenedIn(PlayerEditor? player) {
     _currentlyOpenedIn.remove(player);
   }
 }
@@ -28,15 +28,14 @@ enum AudioType {
 String audioTypeDescription(AudioType audioType) {
   switch (audioType) {
     case AudioType.network:
-      return "network";
+      return 'network';
     case AudioType.liveStream:
-      return "liveStream";
+      return 'liveStream';
     case AudioType.file:
-      return "file";
+      return 'file';
     case AudioType.asset:
-      return "asset";
+      return 'asset';
   }
-  return null;
 }
 
 enum ImageType {
@@ -48,20 +47,25 @@ enum ImageType {
 String imageTypeDescription(ImageType imageType) {
   switch (imageType) {
     case ImageType.network:
-      return "network";
+      return 'network';
     case ImageType.file:
-      return "file";
+      return 'file';
     case ImageType.asset:
-      return "asset";
+      return 'asset';
   }
-  return null;
 }
 
 @immutable
 class MetasImage {
+  const MetasImage({
+    required this.path,
+    required this.type,
+    this.package,
+  });
+
   final String path;
-  final String package;
   final ImageType type;
+  final String? package;
 
   const MetasImage.network(this.path)
       : type = ImageType.network,
@@ -90,13 +94,13 @@ class MetasImage {
 }
 
 class Metas {
-  String id;
-  final String title;
-  final String artist;
-  final String album;
-  final Map<String, dynamic> extra;
-  final MetasImage image;
-  final MetasImage onImageLoadFail;
+  String? id;
+  final String? title;
+  final String? artist;
+  final String? album;
+  final Map<String, dynamic>? extra;
+  final MetasImage? image;
+  final MetasImage? onImageLoadFail;
 
   Metas({
     this.id,
@@ -107,9 +111,7 @@ class Metas {
     this.extra,
     this.onImageLoadFail,
   }) {
-    if (this.id == null) {
-      this.id = Uuid().v4();
-    }
+    id ??= Uuid().v4();
   }
 
   @override
@@ -132,15 +134,15 @@ class Metas {
       onImageLoadFail.hashCode;
 
   Metas copyWith({
-    String id,
-    String title,
-    String artist,
-    String album,
-    Map<String, dynamic> extra,
-    MetasImage image,
-    MetasImage onImageLoadFail,
+    String? id,
+    String? title,
+    String? artist,
+    String? album,
+    Map<String, dynamic>? extra,
+    MetasImage? image,
+    MetasImage? onImageLoadFail,
   }) {
-    return new Metas(
+    return Metas(
       id: id ?? this.id,
       title: title ?? this.title,
       artist: artist ?? this.artist,
@@ -154,69 +156,69 @@ class Metas {
 
 class Audio extends Playable {
   final String path;
-  final String package;
+  final String? package;
   final AudioType audioType;
   Metas _metas;
-  Map<String, dynamic> _networkHeaders;
-  final bool cached; //download audio then play it
-  final double playSpeed;
+  final Map<String, String>? _networkHeaders;
+  final bool? cached; // download audio then play it
+  final double? playSpeed;
 
   Metas get metas => _metas;
 
-  Map<String, dynamic> get networkHeaders => _networkHeaders;
+  Map<String, String>? get networkHeaders => _networkHeaders;
 
   Audio._({
-    this.path,
+    required this.path,
+    required this.audioType,
     this.package,
-    this.audioType,
     this.cached,
     this.playSpeed,
-    Map<String, dynamic> headers,
-    Metas metas,
-  })  : _metas = metas,
+    Map<String, String>? headers,
+    Metas? metas,
+  })  : _metas = metas ?? Metas(),
         _networkHeaders = headers;
 
   Audio(
     this.path, {
-    Metas metas,
+    Metas? metas,
     this.package,
     this.playSpeed,
   })  : audioType = AudioType.asset,
         _networkHeaders = null,
         cached = false,
-        _metas = metas;
+        _metas = metas ?? Metas();
 
   Audio.file(
     this.path, {
-    Metas metas,
+    Metas? metas,
     this.playSpeed,
   })  : audioType = AudioType.file,
         package = null,
         _networkHeaders = null,
         cached = false,
-        _metas = metas;
+        _metas = metas ?? Metas();
 
   Audio.network(
     this.path, {
-    Metas metas,
-    Map<String, dynamic> headers,
+    Metas? metas,
+    Map<String, String>? headers,
     this.cached = false,
     this.playSpeed,
   })  : audioType = AudioType.network,
         package = null,
         _networkHeaders = headers,
-        _metas = metas;
+        _metas = metas ?? Metas();
 
   Audio.liveStream(
     this.path, {
-    Metas metas,
+    Metas? metas,
     this.playSpeed,
-    Map<String, dynamic> headers,
+    Map<String, String>? headers,
   })  : audioType = AudioType.liveStream,
         package = null,
         _networkHeaders = headers,
         cached = false,
-        _metas = metas;
+        _metas = metas ?? Metas();
 
   @override
   bool operator ==(Object other) =>
@@ -245,13 +247,13 @@ class Audio extends Playable {
   }
 
   void updateMetas({
-    String title,
-    String artist,
-    String album,
-    Map<String, dynamic> extra,
-    MetasImage image,
+    String? title,
+    String? artist,
+    String? album,
+    Map<String, dynamic>? extra,
+    MetasImage? image,
   }) {
-    this._metas = (_metas ?? Metas()).copyWith(
+    _metas = _metas.copyWith(
       title: title,
       artist: artist,
       album: album,
@@ -264,20 +266,20 @@ class Audio extends Playable {
   }
 
   Audio copyWith({
-    String path,
-    String package,
-    AudioType audioType,
-    Metas metas,
-    double playSpeed,
-    Map<String, dynamic> headers,
-    bool cached,
+    String? path,
+    String? package,
+    AudioType? audioType,
+    Metas? metas,
+    double? playSpeed,
+    Map<String, String>? headers,
+    bool? cached,
   }) {
     return Audio._(
       path: path ?? this.path,
       package: package ?? this.package,
       audioType: audioType ?? this.audioType,
-      metas: metas ?? this._metas,
-      headers: headers ?? this._networkHeaders,
+      metas: metas ?? _metas,
+      headers: headers ?? _networkHeaders,
       playSpeed: playSpeed ?? this.playSpeed,
       cached: cached ?? this.cached,
     );
@@ -294,12 +296,12 @@ class Playlist extends Playable {
   int get startIndex => _startIndex;
 
   set startIndex(int newValue) {
-    if (newValue < this.audios.length) {
+    if (newValue < audios.length) {
       _startIndex = newValue;
     }
   }
 
-  Playlist({List<Audio> audios, int startIndex = 0}) {
+  Playlist({List<Audio>? audios, int startIndex = 0}) {
     if (audios != null) {
       this.audios.addAll(audios);
     }
@@ -307,38 +309,35 @@ class Playlist extends Playable {
   }
 
   Playlist copyWith({
-    List<Audio> audios,
-    int startIndex,
+    List<Audio>? audios,
+    int? startIndex,
   }) {
-    return new Playlist(
+    return Playlist(
       audios: audios ?? this.audios,
-      startIndex: startIndex ?? this._startIndex,
+      startIndex: startIndex ?? _startIndex,
     );
   }
 
   int get numberOfItems => audios.length;
 
   Playlist add(Audio audio) {
-    if (audio != null) {
-      this.audios.add(audio);
-
-      final index = this.audios.length - 1;
-      super.currentlyOpenedIn.forEach((playerEditor) {
-        playerEditor.onAudioAddedAt(index);
-      });
-    }
+    audios.add(audio);
+    final index = audios.length - 1;
+    super.currentlyOpenedIn.forEach((playerEditor) {
+      playerEditor.onAudioAddedAt(index);
+    });
     return this;
   }
 
   Playlist insert(int index, Audio audio) {
-    if (audio != null && index >= 0) {
-      if (index < this.audios.length) {
-        this.audios.insert(index, audio);
+    if (index >= 0) {
+      if (index < audios.length) {
+        audios.insert(index, audio);
         super.currentlyOpenedIn.forEach((playerEditor) {
           playerEditor.onAudioAddedAt(index);
         });
       } else {
-        return this.add(audio);
+        return add(audio);
       }
     }
     return this;
@@ -346,10 +345,10 @@ class Playlist extends Playable {
 
   Playlist replaceAt(int index, PlaylistAudioReplacer replacer,
       {bool keepPlayingPositionIfCurrent = false}) {
-    if (index < this.audios.length && replacer != null) {
-      final oldElement = this.audios.elementAt(index);
+    if (index < audios.length) {
+      final oldElement = audios.elementAt(index);
       final newElement = replacer(oldElement);
-      this.audios[index] = newElement;
+      audios[index] = newElement;
       super.currentlyOpenedIn.forEach((playerEditor) {
         playerEditor.onAudioReplacedAt(index, keepPlayingPositionIfCurrent);
       });
@@ -358,25 +357,22 @@ class Playlist extends Playable {
   }
 
   Playlist addAll(List<Audio> audios) {
-    if (audios != null) {
-      this.audios.addAll(audios);
-    }
+    this.audios.addAll(audios);
     return this;
   }
 
   bool remove(Audio audio) {
-    if (audio == null) return false;
-    final index = this.audios.indexOf(audio);
-    final bool removed = this.audios.remove(audio);
+    final index = audios.indexOf(audio);
+    final removed = audios.remove(audio);
     super.currentlyOpenedIn.forEach((playerEditor) {
       playerEditor.onAudioRemovedAt(index);
     });
-    //here maybe stop the player if playing this index
+    // here maybe stop the player if playing this index
     return removed;
   }
 
   Audio removeAtIndex(int index) {
-    Audio removedAudio = this.audios.removeAt(index);
+    final removedAudio = audios.removeAt(index);
     super.currentlyOpenedIn.forEach((playerEditor) {
       playerEditor.onAudioRemovedAt(index);
     });
@@ -384,7 +380,7 @@ class Playlist extends Playable {
   }
 
   bool contains(Audio audio) {
-    return this.audios.contains(audio);
+    return audios.contains(audio);
   }
 
   @override
@@ -399,35 +395,34 @@ class Playlist extends Playable {
   int get hashCode => audios.hashCode ^ startIndex.hashCode;
 }
 
-void writeAudioMetasInto(
-    Map<String, dynamic> params, /* nullable */ Metas metas) {
+void writeAudioMetasInto(Map<String, dynamic> params, Metas? metas) {
   if (metas != null) {
-    if (metas.title != null) params["song.title"] = metas.title;
-    if (metas.artist != null) params["song.artist"] = metas.artist;
-    if (metas.album != null) params["song.album"] = metas.album;
+    if (metas.title != null) params['song.title'] = metas.title;
+    if (metas.artist != null) params['song.artist'] = metas.artist;
+    if (metas.album != null) params['song.album'] = metas.album;
     writeAudioImageMetasInto(params, metas.image);
     writeAudioImageMetasInto(params, metas.onImageLoadFail,
-        suffix: ".onLoadFail");
+        suffix: '.onLoadFail');
     if (metas.id != null) {
-      params["song.trackID"] = metas.id;
+      params['song.trackID'] = metas.id;
     }
   }
 }
 
 void writeAudioImageMetasInto(
-    Map<String, dynamic> params, /* nullable */ MetasImage metasImage,
-    {String suffix = ""}) {
+    Map<String, dynamic> params, MetasImage? metasImage,
+    {String suffix = ''}) {
   if (metasImage != null) {
-    params["song.image$suffix"] = metasImage.path;
-    params["song.imageType$suffix"] = imageTypeDescription(metasImage.type);
-    params.addIfNotNull("song.imagePackage$suffix", metasImage.package);
+    params['song.image$suffix'] = metasImage.path;
+    params['song.imageType$suffix'] = imageTypeDescription(metasImage.type);
+    params.addIfNotNull('song.imagePackage$suffix', metasImage.package);
   }
 }
 
 class PlayerGroupMetas {
-  final String title;
-  final String subTitle;
-  final MetasImage image;
+  final String? title;
+  final String? subTitle;
+  final MetasImage? image;
 
   PlayerGroupMetas({
     this.title,
