@@ -1,7 +1,6 @@
 import 'dart:async';
 
 import 'package:assets_audio_player/assets_audio_player.dart';
-import 'package:flutter/cupertino.dart';
 import 'package:flutter/services.dart';
 import 'package:rxdart/rxdart.dart';
 
@@ -32,19 +31,19 @@ class AssetsAudioPlayerGroup {
 
   final PlayInBackground playInBackground;
 
-  AssetsAudioPlayerGroupErrorHandler onErrorDo; //custom error Handler
+  AssetsAudioPlayerGroupErrorHandler? onErrorDo; // custom error Handler
 
   final PlayerGroupMetasCallback updateNotification;
 
-  final PlayerGroupCallback onNotificationOpened;
+  final PlayerGroupCallback? onNotificationOpened;
 
-  final PlayerGroupCallback onNotificationPlay;
-  final PlayerGroupCallback onNotificationPause;
-  final PlayerGroupCallback onNotificationStop;
+  final PlayerGroupCallback? onNotificationPlay;
+  final PlayerGroupCallback? onNotificationPause;
+  final PlayerGroupCallback? onNotificationStop;
 
   final Map<Audio, AssetsAudioPlayer> _audiosWithPlayers = {};
 
-  //copy of _audiosWithPlayers
+  // copy of _audiosWithPlayers
   Map<Audio, AssetsAudioPlayer> get audiosWithPlayers =>
       Map.from(_audiosWithPlayers);
 
@@ -57,11 +56,13 @@ class AssetsAudioPlayerGroup {
 
   ValueStream<bool> get isPlaying => _isPlaying.stream;
 
-  //TODO add streams for audio finished
+  NotificationSettings? __notificationSettings;
+
+  // TODO add streams for audio finished
 
   AssetsAudioPlayerGroup({
     this.showNotification = _DEFAULT_SHOW_NOTIFICATION,
-    @required this.updateNotification,
+    required this.updateNotification,
     this.notificationStopEnabled = _DEFAULT_NOTIFICATION_STOP_ENABLED,
     this.onNotificationOpened,
     this.onNotificationPlay,
@@ -70,17 +71,17 @@ class AssetsAudioPlayerGroup {
     this.respectSilentMode = _DEFAULT_RESPECT_SILENT_MODE,
     this.playInBackground = _DEFAULT_PLAY_IN_BACKGROUND,
   }) {
-    //default action, can be overriden using player.onErrorDo = (error, player) { ACTION };
-    this.onErrorDo = (group, errorHandler) {
+    // default action, can be overriden using player.onErrorDo = (error, player) { ACTION };
+    onErrorDo = (group, errorHandler) {
       print(errorHandler.error.message);
       errorHandler.player.stop();
     };
   }
 
   List<PlayingAudio> get playingAudios {
-    final List<PlayingAudio> audios = <PlayingAudio>[];
+    final audios = <PlayingAudio>[];
     for (final player in players) {
-      final audio = player.current?.value?.audio;
+      final audio = player.current.value?.audio;
       if (audio != null) {
         audios.add(audio);
       }
@@ -88,57 +89,53 @@ class AssetsAudioPlayerGroup {
     return audios;
   }
 
-  NotificationSettings __notificationSettings;
-
-  //intialize the first time
+  // intialize the first time
   NotificationSettings get _notificationSettings {
-    if (__notificationSettings == null) {
-      __notificationSettings = NotificationSettings(
-          stopEnabled: this.notificationStopEnabled,
-          seekBarEnabled: false,
-          nextEnabled: false,
-          prevEnabled: false,
-          customPlayPauseAction: (player) {
-            if (player.isPlaying.value) {
-              if (this.onNotificationPause != null) {
-                this.onNotificationPause(this, playingAudios);
-              } else {
-                _pause();
-              }
+    __notificationSettings ??= NotificationSettings(
+        stopEnabled: notificationStopEnabled,
+        seekBarEnabled: false,
+        nextEnabled: false,
+        prevEnabled: false,
+        customPlayPauseAction: (player) {
+          if (player.isPlaying.value == true) {
+            if (onNotificationPause != null) {
+              onNotificationPause!(this, playingAudios);
             } else {
-              if (this.onNotificationPlay != null) {
-                this.onNotificationPlay(this, playingAudios);
-              } else {
-                _play();
-              }
+              _pause();
             }
-          },
-          customStopAction: (player) {
-            if (this.onNotificationStop != null) {
-              this.onNotificationStop(this, playingAudios);
+          } else {
+            if (onNotificationPlay != null) {
+              onNotificationPlay!(this, playingAudios);
             } else {
-              _stop();
+              _play();
             }
-          });
-    }
-    return __notificationSettings;
+          }
+        },
+        customStopAction: (player) {
+          if (onNotificationStop != null) {
+            onNotificationStop!(this, playingAudios);
+          } else {
+            _stop();
+          }
+        });
+    return __notificationSettings!;
   }
 
   Future<void> add(
     Audio audio, {
     LoopMode loopMode = LoopMode.none,
-    double volume,
-    Duration seek,
-    double playSpeed,
+    double? volume,
+    Duration? seek,
+    double? playSpeed,
   }) async {
     final player = AssetsAudioPlayer.newPlayer();
-    player.open(
+    await player.open(
       audio,
       showNotification: false,
-      //not need here, we'll call another method `changeNotificationForGroup`
+      // not need here, we'll call another method `changeNotificationForGroup`
       seek: seek,
-      autoStart: isPlaying.value,
-      //need to play() for player group
+      autoStart: isPlaying.value ?? false,
+      // need to play() for player group
       volume: volume,
       loopMode: loopMode,
       respectSilentMode: respectSilentMode,
@@ -150,7 +147,9 @@ class AssetsAudioPlayerGroup {
   }
 
   Future<void> addAll(List<Audio> audios) async {
-    for (Audio audio in audios) await add(audio);
+    for (final audio in audios) {
+      await add(audio);
+    }
   }
 
   Future<void> removeAudio(Audio audio) async {
@@ -164,7 +163,7 @@ class AssetsAudioPlayerGroup {
   }
 
   Future<void> _addPlayer(Audio audio, AssetsAudioPlayer player) async {
-    StreamSubscription finishedSubscription;
+    late StreamSubscription finishedSubscription;
     finishedSubscription = player.playlistFinished.listen((finished) {
       if (finished) {
         finishedSubscription.cancel();
@@ -183,72 +182,67 @@ class AssetsAudioPlayerGroup {
   }
 
   void _onPlayerError(ErrorHandler errorHandler) {
-    if (this.onErrorDo != null) {
-      this.onErrorDo(this, errorHandler);
+    if (onErrorDo != null) {
+      onErrorDo!(this, errorHandler);
     }
   }
 
-  ///Called when an audio is added or removed (/finished)
+  /// Called when an audio is added or removed (/finished)
   Future<void> _onPlayersChanged() async {
-    if (updateNotification != null) {
-      final bool isPlaying = this.isPlaying.value;
-      final newNotificationsMetas =
-          await updateNotification(this, playingAudios);
+    final isPlaying = this.isPlaying.value ?? false;
+    final newNotificationsMetas = await updateNotification(this, playingAudios);
 
-      String firstPlayerId;
-      if (audios.isNotEmpty) {
-        firstPlayerId = players.first?.id;
-      }
-
-      changeNotificationForGroup(
-        //TODO find a way to protect it
-        this,
-        isPlaying: isPlaying,
-        firstPlayerId: firstPlayerId,
-        display: playingAudios.isNotEmpty,
-        notificationSettings: this._notificationSettings,
-        metas: Metas(
-          title: newNotificationsMetas.title,
-          artist: newNotificationsMetas.subTitle,
-          image: newNotificationsMetas.image,
-        ),
-      );
+    String? firstPlayerId;
+    if (audios.isNotEmpty) {
+      firstPlayerId = players.first.id;
     }
+
+    await changeNotificationForGroup(
+      // TODO find a way to protect it
+      this,
+      isPlaying: isPlaying,
+      firstPlayerId: firstPlayerId,
+      display: playingAudios.isNotEmpty,
+      notificationSettings: _notificationSettings,
+      metas: Metas(
+        title: newNotificationsMetas.title,
+        artist: newNotificationsMetas.subTitle,
+        image: newNotificationsMetas.image,
+      ),
+    );
   }
 
   Future<void> changeNotificationForGroup(
     AssetsAudioPlayerGroup playerGroup, {
-    Metas metas,
-    bool display,
-    String firstPlayerId,
-    NotificationSettings notificationSettings,
+    required NotificationSettings notificationSettings,
+    Metas? metas,
+    bool? display,
+    String? firstPlayerId,
     bool isPlaying = true,
   }) async {
-    if (playerGroup != null) {
-      final Map<String, dynamic> params = {
-        "id": firstPlayerId,
-        "isPlaying": isPlaying,
-        "display": display,
-      };
+    final params = {
+      'id': firstPlayerId,
+      'isPlaying': isPlaying,
+      'display': display,
+    };
 
-      writeAudioMetasInto(params, metas);
-      writeNotificationSettingsInto(params, notificationSettings);
+    writeAudioMetasInto(params, metas);
+    writeNotificationSettingsInto(params, notificationSettings);
 
-      await _sendChannel.invokeMethod('forceNotificationForGroup', params);
-    }
+    await _sendChannel.invokeMethod('forceNotificationForGroup', params);
   }
 
   Future<void> play() {
     return _play();
   }
 
-  Future<void> _play({AssetsAudioPlayer except}) async {
-    for (AssetsAudioPlayer player in players) {
+  Future<void> _play({AssetsAudioPlayer? except}) async {
+    for (final player in players) {
       if (player != except) {
         await player.play();
       }
     }
-    _isPlaying.value = true;
+    _isPlaying.add(true);
     await _onPlayersChanged();
   }
 
@@ -256,13 +250,13 @@ class AssetsAudioPlayerGroup {
     return _pause();
   }
 
-  Future<void> _pause({AssetsAudioPlayer except}) async {
-    for (AssetsAudioPlayer player in players) {
+  Future<void> _pause({AssetsAudioPlayer? except}) async {
+    for (final player in players) {
       if (player != except) {
         await player.pause();
       }
     }
-    _isPlaying.value = false;
+    _isPlaying.add(false);
     await _onPlayersChanged();
   }
 
@@ -270,15 +264,15 @@ class AssetsAudioPlayerGroup {
     return _stop();
   }
 
-  Future<void> _stop({AssetsAudioPlayer except}) async {
+  Future<void> _stop({AssetsAudioPlayer? except}) async {
     //copy _players because _stop remove the player from the list
-    final List<AssetsAudioPlayer> copyList = List.from(players);
+    final copyList = List.from(players);
     for (AssetsAudioPlayer player in copyList) {
       if (player != except) {
         await player.stop();
       }
     }
-    _isPlaying.value = false;
+    _isPlaying.add(false);
     await _onPlayersChanged();
   }
 
@@ -295,7 +289,7 @@ class AssetsAudioPlayerGroup {
   }
 
   Future<void> playOrPause() async {
-    if (isPlaying.value) {
+    if (isPlaying.value == true) {
       await pause();
     } else {
       await play();
