@@ -24,10 +24,9 @@ class AssetAudioPlayerIcons {
 
   static const _kFontFam = 'AssetAudioPlayer';
 
-  static const IconData play = const IconData(0xe800, fontFamily: _kFontFam);
-  static const IconData stop = const IconData(0xe801, fontFamily: _kFontFam);
-  static const IconData pause = const IconData(0xe802, fontFamily: _kFontFam);
-  static const IconData to_end = const IconData(0xe803, fontFamily: _kFontFam);
-  static const IconData to_start =
-      const IconData(0xe804, fontFamily: _kFontFam);
+  static const IconData play = IconData(0xe800, fontFamily: _kFontFam);
+  static const IconData stop = IconData(0xe801, fontFamily: _kFontFam);
+  static const IconData pause = IconData(0xe802, fontFamily: _kFontFam);
+  static const IconData to_end = IconData(0xe803, fontFamily: _kFontFam);
+  static const IconData to_start = IconData(0xe804, fontFamily: _kFontFam);
 }
