@@ -2,6 +2,8 @@
 // Generated file. Do not edit.
 //
 
+// ignore_for_file: lines_longer_than_80_chars
+
 import 'package:assets_audio_player_web/web/assets_audio_player_web.dart';
 
 import 'package:flutter_web_plugins/flutter_web_plugins.dart';
@@ -9,5 +11,5 @@ import 'package:flutter_web_plugins/flutter_web_plugins.dart';
 // ignore: public_member_api_docs
 void registerPlugins(Registrar registrar) {
   AssetsAudioPlayerWebPlugin.registerWith(registrar);
-  //registrar.registerMessageHandler();
+  registrar.registerMessageHandler();
 }
