@@ -26,66 +26,72 @@ class MyApp extends StatefulWidget {
 class _MyAppState extends State<MyApp> {
   final audios = <Audio>[
     Audio.network(
-      "https://files.freemusicarchive.org/storage-freemusicarchive-org/music/Music_for_Video/springtide/Sounds_strange_weird_but_unmistakably_romantic_Vol1/springtide_-_03_-_We_Are_Heading_to_the_East.mp3",
+      'https://files.freemusicarchive.org/storage-freemusicarchive-org/music/Music_for_Video/springtide/Sounds_strange_weird_but_unmistakably_romantic_Vol1/springtide_-_03_-_We_Are_Heading_to_the_East.mp3',
       metas: Metas(
-        title: "Online",
-        artist: "Florent Champigny",
-        album: "OnlineAlbum",
-        image: MetasImage.network("https://image.shutterstock.com/image-vector/pop-music-text-art-colorful-600w-515538502.jpg"),
+        title: 'Online',
+        artist: 'Florent Champigny',
+        album: 'OnlineAlbum',
+        image: MetasImage.network(
+            'https://image.shutterstock.com/image-vector/pop-music-text-art-colorful-600w-515538502.jpg'),
       ),
     ),
     Audio(
-      "assets/audios/rock.mp3",
+      'assets/audios/rock.mp3',
       metas: Metas(
-        title: "Rock",
-        artist: "Florent Champigny",
-        album: "RockAlbum",
-        image: MetasImage.network("https://static.radio.fr/images/broadcasts/cb/ef/2075/c300.png"),
+        title: 'Rock',
+        artist: 'Florent Champigny',
+        album: 'RockAlbum',
+        image: MetasImage.network(
+            'https://static.radio.fr/images/broadcasts/cb/ef/2075/c300.png'),
       ),
     ),
     Audio(
-      "assets/audios/country.mp3",
+      'assets/audios/country.mp3',
       metas: Metas(
-        title: "Country",
-        artist: "Florent Champigny",
-        album: "CountryAlbum",
-        image: MetasImage.asset("assets/images/country.jpg"),
+        title: 'Country',
+        artist: 'Florent Champigny',
+        album: 'CountryAlbum',
+        image: MetasImage.asset('assets/images/country.jpg'),
       ),
     ),
     Audio(
-      "assets/audios/electronic.mp3",
+      'assets/audios/electronic.mp3',
       metas: Metas(
-        title: "Electronic",
-        artist: "Florent Champigny",
-        album: "ElectronicAlbum",
-        image: MetasImage.network("https://i.ytimg.com/vi/nVZNy0ybegI/maxresdefault.jpg"),
+        title: 'Electronic',
+        artist: 'Florent Champigny',
+        album: 'ElectronicAlbum',
+        image: MetasImage.network(
+            'https://i.ytimg.com/vi/nVZNy0ybegI/maxresdefault.jpg'),
       ),
     ),
     Audio(
-      "assets/audios/hiphop.mp3",
+      'assets/audios/hiphop.mp3',
       metas: Metas(
-        title: "HipHop",
-        artist: "Florent Champigny",
-        album: "HipHopAlbum",
-        image: MetasImage.network("https://beyoudancestudio.ch/wp-content/uploads/2019/01/apprendre-danser.hiphop-1.jpg"),
+        title: 'HipHop',
+        artist: 'Florent Champigny',
+        album: 'HipHopAlbum',
+        image: MetasImage.network(
+            'https://beyoudancestudio.ch/wp-content/uploads/2019/01/apprendre-danser.hiphop-1.jpg'),
       ),
     ),
     Audio(
-      "assets/audios/pop.mp3",
+      'assets/audios/pop.mp3',
       metas: Metas(
-        title: "Pop",
-        artist: "Florent Champigny",
-        album: "PopAlbum",
-        image: MetasImage.network("https://image.shutterstock.com/image-vector/pop-music-text-art-colorful-600w-515538502.jpg"),
+        title: 'Pop',
+        artist: 'Florent Champigny',
+        album: 'PopAlbum',
+        image: MetasImage.network(
+            'https://image.shutterstock.com/image-vector/pop-music-text-art-colorful-600w-515538502.jpg'),
       ),
     ),
     Audio(
-      "assets/audios/instrumental.mp3",
+      'assets/audios/instrumental.mp3',
       metas: Metas(
-        title: "Instrumental",
-        artist: "Florent Champigny",
-        album: "InstrumentalAlbum",
-        image: MetasImage.network("https://i.ytimg.com/vi/zv_0dSfknBc/maxresdefault.jpg"),
+        title: 'Instrumental',
+        artist: 'Florent Champigny',
+        album: 'InstrumentalAlbum',
+        image: MetasImage.network(
+            'https://i.ytimg.com/vi/zv_0dSfknBc/maxresdefault.jpg'),
       ),
     ),
   ];
@@ -96,16 +102,16 @@ class _MyAppState extends State<MyApp> {
   @override
   void initState() {
     _subscriptions.add(_assetsAudioPlayer.playlistFinished.listen((data) {
-      print("finished : $data");
+      print('finished : $data');
     }));
     _subscriptions.add(_assetsAudioPlayer.playlistAudioFinished.listen((data) {
-      print("playlistAudioFinished : $data");
+      print('playlistAudioFinished : $data');
     }));
     _subscriptions.add(_assetsAudioPlayer.current.listen((data) {
-      print("current : $data");
+      print('current : $data');
     }));
     _subscriptions.add(_assetsAudioPlayer.onReadyToPlay.listen((audio) {
-      print("onRedayToPlay : $audio");
+      print('onRedayToPlay : $audio');
     }));
     super.initState();
   }
@@ -139,9 +145,10 @@ class _MyAppState extends State<MyApp> {
                   fit: StackFit.passthrough,
                   children: <Widget>[
                     _assetsAudioPlayer.builderCurrent(
-                      builder: (BuildContext context, Playing playing) {
+                      builder: (BuildContext context, Playing? playing) {
                         if (playing != null) {
-                          final myAudio = find(this.audios, playing.audio.assetAudioPath);
+                          final myAudio =
+                              find(audios, playing.audio.assetAudioPath);
                           return Padding(
                             padding: const EdgeInsets.all(8.0),
                             child: Neumorphic(
@@ -151,19 +158,22 @@ class _MyAppState extends State<MyApp> {
                                 surfaceIntensity: 1,
                                 shape: NeumorphicShape.concave,
                               ),
-                              child: myAudio.metas.image.type == ImageType.network
-                                  ? Image.network(
-                                      myAudio.metas.image.path,
-                                      height: 150,
-                                      width: 150,
-                                      fit: BoxFit.contain,
-                                    )
-                                  : Image.asset(
-                                      myAudio.metas.image.path,
-                                      height: 150,
-                                      width: 150,
-                                      fit: BoxFit.contain,
-                                    ),
+                              child: myAudio.metas.image?.path == null
+                                  ? const SizedBox()
+                                  : myAudio.metas.image?.type ==
+                                          ImageType.network
+                                      ? Image.network(
+                                          myAudio.metas.image!.path,
+                                          height: 150,
+                                          width: 150,
+                                          fit: BoxFit.contain,
+                                        )
+                                      : Image.asset(
+                                          myAudio.metas.image!.path,
+                                          height: 150,
+                                          width: 150,
+                                          fit: BoxFit.contain,
+                                        ),
                             ),
                           );
                         }
@@ -179,7 +189,8 @@ class _MyAppState extends State<MyApp> {
                         padding: EdgeInsets.all(18),
                         margin: EdgeInsets.all(18),
                         onPressed: () {
-                          AssetsAudioPlayer.playAndForget(Audio("assets/audios/horn.mp3"));
+                          AssetsAudioPlayer.playAndForget(
+                              Audio('assets/audios/horn.mp3'));
                         },
                         child: Icon(
                           Icons.add_alert,
@@ -195,7 +206,8 @@ class _MyAppState extends State<MyApp> {
                 SizedBox(
                   height: 20,
                 ),
-                _assetsAudioPlayer.builderCurrent(builder: (context, playing) {
+                _assetsAudioPlayer.builderCurrent(
+                    builder: (context, Playing? playing) {
                   if (playing == null) {
                     return SizedBox();
                   }
@@ -227,11 +239,12 @@ class _MyAppState extends State<MyApp> {
                               });
                         },
                       ),
-                      _assetsAudioPlayer.builderRealtimePlayingInfos(builder: (context, infos) {
+                      _assetsAudioPlayer.builderRealtimePlayingInfos(
+                          builder: (context, RealtimePlayingInfos? infos) {
                         if (infos == null) {
                           return SizedBox();
                         }
-                        //print("infos: $infos");
+                        //print('infos: $infos');
                         return Column(
                           children: [
                             PositionSeekWidget(
@@ -245,19 +258,21 @@ class _MyAppState extends State<MyApp> {
                               mainAxisAlignment: MainAxisAlignment.center,
                               children: [
                                 NeumorphicButton(
-                                  child: Text("-10"),
                                   onPressed: () {
-                                    _assetsAudioPlayer.seekBy(Duration(seconds: -10));
+                                    _assetsAudioPlayer
+                                        .seekBy(Duration(seconds: -10));
                                   },
+                                  child: Text('-10'),
                                 ),
                                 SizedBox(
                                   width: 12,
                                 ),
                                 NeumorphicButton(
-                                  child: Text("+10"),
                                   onPressed: () {
-                                    _assetsAudioPlayer.seekBy(Duration(seconds: 10));
+                                    _assetsAudioPlayer
+                                        .seekBy(Duration(seconds: 10));
                                   },
+                                  child: Text('+10'),
                                 ),
                               ],
                             )
@@ -271,9 +286,10 @@ class _MyAppState extends State<MyApp> {
                   height: 20,
                 ),
                 Expanded(
-                  child: _assetsAudioPlayer.builderCurrent(builder: (BuildContext context, Playing playing) {
+                  child: _assetsAudioPlayer.builderCurrent(
+                      builder: (BuildContext context, Playing playing) {
                     return SongsSelector(
-                      audios: this.audios,
+                      audios: audios,
                       onPlaylistSelected: (myAudios) {
                         _assetsAudioPlayer.open(
                           Playlist(audios: myAudios),
