@@ -22,71 +22,73 @@ class MyApp extends StatefulWidget {
 class _MyAppState extends State<MyApp> {
   final audios = <Audio>[
     Audio.network(
-      "https://files.freemusicarchive.org/storage-freemusicarchive-org/music/Music_for_Video/springtide/Sounds_strange_weird_but_unmistakably_romantic_Vol1/springtide_-_03_-_We_Are_Heading_to_the_East.mp3",
+      'https://files.freemusicarchive.org/storage-freemusicarchive-org/music/Music_for_Video/springtide/Sounds_strange_weird_but_unmistakably_romantic_Vol1/springtide_-_03_-_We_Are_Heading_to_the_East.mp3',
       metas: Metas(
-        title: "Online",
-        artist: "Florent Champigny",
-        album: "OnlineAlbum",
-        image: MetasImage.network("https://image.shutterstock.com/image-vector/pop-music-text-art-colorful-600w-515538502.jpg"),
+        title: 'Online',
+        artist: 'Florent Champigny',
+        album: 'OnlineAlbum',
+        image: MetasImage.network(
+            'https://image.shutterstock.com/image-vector/pop-music-text-art-colorful-600w-515538502.jpg'),
       ),
     ),
     Audio(
-      "assets/audios/rock.mp3",
+      'assets/audios/rock.mp3',
       metas: Metas(
-        title: "Rock",
-        artist: "Florent Champigny",
-        album: "RockAlbum",
-        image: MetasImage.network("https://static.radio.fr/images/broadcasts/cb/ef/2075/c300.png"),
+        title: 'Rock',
+        artist: 'Florent Champigny',
+        album: 'RockAlbum',
+        image: MetasImage.network(
+            'https://static.radio.fr/images/broadcasts/cb/ef/2075/c300.png'),
       ),
     ),
     Audio(
-      "assets/audios/country.mp3",
+      'assets/audios/country.mp3',
       metas: Metas(
-        title: "Country",
-        artist: "Florent Champigny",
-        album: "CountryAlbum",
-        image: MetasImage.asset("assets/images/country.jpg"),
+        title: 'Country',
+        artist: 'Florent Champigny',
+        album: 'CountryAlbum',
+        image: MetasImage.asset('assets/images/country.jpg'),
       ),
     ),
     Audio(
-      "assets/audios/electronic.mp3",
+      'assets/audios/electronic.mp3',
       metas: Metas(
-        title: "Electronic",
-        artist: "Florent Champigny",
-        album: "ElectronicAlbum",
-        image: MetasImage.network("https://i.ytimg.com/vi/nVZNy0ybegI/maxresdefault.jpg"),
+        title: 'Electronic',
+        artist: 'Florent Champigny',
+        album: 'ElectronicAlbum',
+        image: MetasImage.network(
+            'https://i.ytimg.com/vi/nVZNy0ybegI/maxresdefault.jpg'),
       ),
     ),
   ];
 
-  final AssetsAudioPlayerGroup _assetsAudioPlayerGroup = AssetsAudioPlayerGroup(
-    updateNotification: (player, playing) async {
-      return PlayerGroupMetas(
-        title: "title",
-        subTitle: "subtitle ${playing.length}",
-        image: MetasImage.asset("assets/images/country.jpg"),
-      );
-    }
-  );
+  final AssetsAudioPlayerGroup _assetsAudioPlayerGroup =
+      AssetsAudioPlayerGroup(updateNotification: (player, playing) async {
+    return PlayerGroupMetas(
+      title: 'title',
+      subTitle: 'subtitle ${playing.length}',
+      image: MetasImage.asset('assets/images/country.jpg'),
+    );
+  });
   //final List<StreamSubscription> _subscriptions = [];
 
   @override
   void initState() {
     /*
     _subscriptions.add(_assetsAudioPlayer.playlistFinished.listen((data) {
-      print("finished : $data");
+      print('finished : $data');
     }));
     _subscriptions.add(_assetsAudioPlayer.playlistAudioFinished.listen((data) {
-      print("playlistAudioFinished : $data");
+      print('playlistAudioFinished : $data');
     }));
     _subscriptions.add(_assetsAudioPlayer.current.listen((data) {
-      print("current : $data");
+      print('current : $data');
     }));
     _subscriptions.add(_assetsAudioPlayer.onReadyToPlay.listen((audio) {
-      print("onRedayToPlay : $audio");
+      print('onRedayToPlay : $audio');
     }));
     _subscriptions.add(_assetsAudioPlayer.playerState.listen((playerState) {
-      print("playerState : $playerState");
+      print('playerState : $playerState');
     }));
     */
     _assetsAudioPlayerGroup.addAll(audios);
@@ -119,7 +121,8 @@ class _MyAppState extends State<MyApp> {
                   SizedBox(
                     height: 20,
                   ),
-                  _assetsAudioPlayerGroup.builderIsPlaying(builder: (context, isPlaying) {
+                  _assetsAudioPlayerGroup.builderIsPlaying(
+                      builder: (context, bool? isPlaying) {
                     if (isPlaying == null) {
                       return SizedBox();
                     }
