@@ -16,20 +16,22 @@ class _MyAppState extends State {
   @override
   void initState() {
     _assetsAudioPlayer.playlistFinished.listen((data) {
-      print("finished : $data");
+      print('finished : $data');
     });
     _assetsAudioPlayer.playlistAudioFinished.listen((data) {
-      print("playlistAudioFinished : $data");
+      print('playlistAudioFinished : $data');
     });
     _assetsAudioPlayer.current.listen((data) {
-      print("current : $data");
+      print('current : $data');
     });
     _assetsAudioPlayer.onReadyToPlay.listen((audio) {
-      print("onReadyToPlay : $audio");
+      print('onReadyToPlay : $audio');
     });
-    _assetsAudioPlayer.open(Audio(
-      "assets/audios/water.mp3",
-    ), loopMode: LoopMode.playlist);
+    _assetsAudioPlayer.open(
+        Audio(
+          'assets/audios/water.mp3',
+        ),
+        loopMode: LoopMode.playlist);
     super.initState();
   }
 
@@ -40,9 +42,8 @@ class _MyAppState extends State {
         body: Center(
           child: _assetsAudioPlayer.builderCurrentPosition(
               builder: (BuildContext context, Duration position) {
-                return Text(position.toString());
-              }
-          ),
+            return Text(position.toString());
+          }),
         ),
       ),
     );
