@@ -7,11 +7,11 @@ import 'cache_downloader.dart';
 AssetsAudioPlayerCacheManager _instance = AssetsAudioPlayerCacheManager._();
 
 class AssetsAudioPlayerCacheManager {
-  //singleton
+  // singleton
   AssetsAudioPlayerCacheManager._();
   factory AssetsAudioPlayerCacheManager() => _instance;
 
-  Map<String, CacheDownloader> _downloadingElements = Map();
+  final Map<String, CacheDownloader> _downloadingElements = {};
 
   Future<Audio> transform(
     AssetsAudioPlayerCache cache,
@@ -22,8 +22,8 @@ class AssetsAudioPlayerCacheManager {
       return audio;
     }
 
-    final String key = await cache.audioKeyTransformer(audio);
-    final String path = await cache.cachePathProvider(audio, key);
+    final key = await cache.audioKeyTransformer(audio);
+    final path = await cache.cachePathProvider(audio, key);
     if (await _fileExists(path)) {
       return audio.copyWith(path: path, audioType: AudioType.file);
     } else {
@@ -31,15 +31,15 @@ class AssetsAudioPlayerCacheManager {
         await _download(audio, path, cacheDownloadListener);
         return audio.copyWith(path: path, audioType: AudioType.file);
       } catch (t) {
-        //TODO
+        // TODO
       }
     }
 
-    return audio; //do not change anything if error
+    return audio; // do not change anything if error
   }
 
   Future<bool> _fileExists(String path) async {
-    final File file = File(path);
+    final file = File(path);
     return await file.exists();
   }
 
@@ -50,26 +50,26 @@ class AssetsAudioPlayerCacheManager {
   ) async {
     print(intoPath);
     if (_downloadingElements.containsKey(intoPath)) {
-      //is already downloading it
+      // is already downloading it
       final downloader = _downloadingElements[intoPath];
-      await downloader.wait(cacheDownloadListener);
+      await downloader?.wait(cacheDownloadListener);
     } else {
       try {
         final downloader = CacheDownloader();
         _downloadingElements[intoPath] = downloader;
-        downloader.downloadAndSave(
+        await downloader.downloadAndSave(
           url: audio.path,
           savePath: intoPath,
-          headers: audio.networkHeaders ?? {},
+          headers: audio.networkHeaders,
         );
         await downloader.wait(cacheDownloadListener);
-        //download finished
+        // download finished
         _downloadingElements.remove(intoPath);
       } catch (t) {
-        //on error, remove also the downloader
+        // on error, remove also the downloader
         _downloadingElements.remove(intoPath);
-        //then throw again the exception
-        throw t;
+        // then throw again the exception
+        rethrow;
       }
     }
   }
