@@ -1,5 +1,5 @@
 class NetworkSettings {
-  final Map<String, dynamic> defaultHeaders = Map();
+  final Map<String, dynamic> defaultHeaders = {};
 
   NetworkSettings();
 }
